vm: split llm/api.rs (3598 lines) into focused submodules
The single-file LLM transport layer mixed call options, request
serialization, response parsing, provider auth, error classification,
context-window resolution, and thinking-block streaming. Split it into
llm/api/ with one leaf module per concern so the wire-format layer stays
comprehensible as more providers land.

- auth.rs, ollama.rs, thinking.rs, errors.rs: leaf helpers + tests
- options.rs: LlmCallOptions, LlmRequestPayload, tool-search configs
- result.rs: LlmResult + vm_build_llm_result
- openai_normalize.rs: message-shape helpers
- response.rs: non-streaming parser + cache-token extraction + tests
- context_window.rs: discovery cache + auto-compact adaptation
- completion.rs: completion dispatch (openai/ollama/fallback)
- transport.rs: SSE + NDJSON streaming + provider dispatch

api.rs shrinks from 3598 to 761 lines (entry points, re-exports, and
transport integration tests). All files are under the 1200-line ceiling
from harn#42. Provider behavior is unchanged: every external `api::*`
import path still resolves via re-exports. `make all` passes end-to-end.

Closes #42
