+++ crates/harn-vm/src/llm/api.rs
@@ -1,4 +1,21 @@
-use std::collections::BTreeMap;
+//! LLM API entry points and re-exports. The transport layer, request/
+//! response parsing, auth, context-window discovery, and option/result
+//! types each live in their own submodule under [`self`]; this file only
+//! wires them together and hosts the `vm_call_llm_full*` chat entry
+//! points that provider-specific completion / agent paths dispatch into.
+
+mod auth;
+mod completion;
+mod context_window;
+mod errors;
+mod ollama;
+mod openai_normalize;
+mod options;
+mod response;
+mod result;
+mod thinking;
+mod transport;
+
 use std::rc::Rc;
 
 use crate::value::{VmError, VmValue};
@@ -6,522 +23,25 @@ use crate::value::{VmError, VmValue};
 use super::mock::{
     fixture_hash, get_replay_mode, load_fixture, mock_llm_response, save_fixture, LlmReplayMode,
 };
-use super::provider::LlmProvider;
 
-/// Sender for streaming text deltas from an in-flight LLM call.
-pub(crate) type DeltaSender = tokio::sync::mpsc::UnboundedSender<String>;
+// ─── Public surface (crate-wide) ────────────────────────────────────────
 
-/// Extended thinking configuration.
-#[derive(Clone, Debug)]
-pub(crate) enum ThinkingConfig {
-    /// Enable with provider defaults.
-    Enabled,
-    /// Enable with a specific token budget.
-    WithBudget(i64),
-}
+pub(crate) use auth::apply_auth_headers;
+pub(crate) use completion::vm_call_completion_full;
+pub(crate) use context_window::adapt_auto_compact_to_provider;
+pub use context_window::fetch_provider_max_context;
+pub(crate) use ollama::{ollama_keep_alive_override, ollama_num_ctx_override};
+pub(crate) use openai_normalize::{debug_log_message_shapes, normalize_openai_style_messages};
+pub(crate) use options::{
+    DeltaSender, LlmCallOptions, LlmRequestPayload, ThinkingConfig, ToolSearchConfig,
+    ToolSearchMode, ToolSearchStrategy, ToolSearchVariant,
+};
+pub(crate) use result::{vm_build_llm_result, LlmResult};
+pub(crate) use transport::vm_call_llm_api_with_body;
 
-/// Which tool-search variant to use. Two shapes today, matching the two
-/// Anthropic variants (also reused as the mental model for the OpenAI path
-/// landing in harn#71). Scripts write the lower-case short name.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum ToolSearchVariant {
-    /// BM25 / natural-language queries. Default when the user wrote just
-    /// `tool_search: true` or omitted the variant.
-    Bm25,
-    /// Python-regex queries (more precise, less ergonomic).
-    Regex,
-}
+use transport::vm_call_llm_api;
 
-impl ToolSearchVariant {
-    pub(crate) fn as_short(self) -> &'static str {
-        match self {
-            ToolSearchVariant::Bm25 => "bm25",
-            ToolSearchVariant::Regex => "regex",
-        }
-    }
-}
-
-/// Implementation of the client-executed tool-search fallback (harn#70).
-/// Only consulted when `ToolSearchMode::Client` resolves (either
-/// explicit or via auto-fallback when the provider lacks native
-/// support). Orthogonal to `ToolSearchVariant`: a user can ask for
-/// `variant: bm25` (the model sees the BM25-style tool) and
-/// `strategy: semantic` (the host runs embedding search under the
-/// hood).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum ToolSearchStrategy {
-    /// In-tree BM25 over the deferred tool corpus. **Default.**
-    Bm25,
-    /// In-tree regex over the deferred tool corpus (case-insensitive).
-    Regex,
-    /// Delegated to the host via the `tool_search/query` bridge RPC so
-    /// integrators can wire embeddings without Harn depending on ML
-    /// crates.
-    Semantic,
-    /// Pure host-side implementation; the VM just round-trips the query
-    /// and promotes whatever names the host returns.
-    Host,
-}
-
-impl ToolSearchStrategy {
-    pub(crate) fn as_short(self) -> &'static str {
-        match self {
-            ToolSearchStrategy::Bm25 => "bm25",
-            ToolSearchStrategy::Regex => "regex",
-            ToolSearchStrategy::Semantic => "semantic",
-            ToolSearchStrategy::Host => "host",
-        }
-    }
-
-    /// Whether this strategy runs entirely inside the VM (no bridge
-    /// hop). Used by the dispatch path to decide between the sync
-    /// in-tree index and the `tool_search/query` RPC.
-    #[allow(dead_code)] // consumed by harness tests + future dispatch refactors
-    pub(crate) fn is_in_tree(self) -> bool {
-        matches!(self, ToolSearchStrategy::Bm25 | ToolSearchStrategy::Regex)
-    }
-
-    /// Map to the in-tree strategy enum used by
-    /// [`crate::llm::tool_search::run_in_tree`]. Panics on non-in-tree
-    /// strategies — callers must gate on `is_in_tree()`.
-    pub(crate) fn as_in_tree(self) -> crate::llm::tool_search::InTreeStrategy {
-        match self {
-            ToolSearchStrategy::Bm25 => crate::llm::tool_search::InTreeStrategy::Bm25,
-            ToolSearchStrategy::Regex => crate::llm::tool_search::InTreeStrategy::Regex,
-            _ => unreachable!("as_in_tree called on {self:?}"),
-        }
-    }
-
-    /// Default strategy for a given variant when the user did not
-    /// specify one explicitly. Native-facing variant leaks into the
-    /// client path as a sensible default: `variant: regex` users
-    /// probably want regex semantics in the fallback too.
-    pub(crate) fn default_for_variant(variant: ToolSearchVariant) -> Self {
-        match variant {
-            ToolSearchVariant::Bm25 => ToolSearchStrategy::Bm25,
-            ToolSearchVariant::Regex => ToolSearchStrategy::Regex,
-        }
-    }
-}
-
-/// How to resolve `tool_search` against the active provider.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum ToolSearchMode {
-    /// Auto-select: native if the provider supports it, client-executed
-    /// fallback otherwise (harn#70). Default.
-    Auto,
-    /// Force the provider's native mechanism; error if unsupported.
-    Native,
-    /// Force client-executed fallback even when native is available.
-    /// Currently errors with a pointer to harn#70 until the fallback lands.
-    Client,
-}
-
-/// User-facing tool_search configuration. Parsed from the `tool_search`
-/// option on `llm_call` / `agent_loop`. Absent means no deferred-loading
-/// machinery is engaged — tools ship eagerly as always.
-#[derive(Clone, Debug)]
-pub(crate) struct ToolSearchConfig {
-    pub variant: ToolSearchVariant,
-    pub mode: ToolSearchMode,
-    /// Tool names that must remain eager even when `defer_loading: true`
-    /// is otherwise set on them. Useful for a "safety net" a skill wants
-    /// always available regardless of the tool-search index's decisions.
-    /// Only consumed by the client-executed path — for the native
-    /// Anthropic path, eagerness is already controlled per-tool via
-    /// `defer_loading`.
-    pub always_loaded: Vec<String>,
-    /// Client-mode implementation strategy. When unset, defaults to
-    /// `ToolSearchStrategy::default_for_variant(variant)`.
-    pub strategy: Option<ToolSearchStrategy>,
-    /// Soft cap on how many deferred tools the client-executed loop
-    /// may promote into the eager set over the life of this call.
-    /// Oldest-promoted tools are evicted when the cap is hit. `None`
-    /// means no cap — rely on the `max_results` per search call.
-    pub budget_tokens: Option<i64>,
-    /// Override for the synthetic tool's name. Default
-    /// `__harn_tool_search`. Lets skills with a brand-specific vocabulary
-    /// name the tool something the model will understand out of the
-    /// box (`find_tool`, `discover_tool`, etc.).
-    pub name: Option<String>,
-    /// When true, the client-mode loop includes a short stub line for
-    /// each deferred tool (name + one-line summary) alongside the
-    /// synthetic search tool so the model knows what's available
-    /// without calling search first. Default: `false` — the Anthropic
-    /// native path also ships no stubs.
-    pub include_stub_listing: bool,
-    /// Canonical native-shape JSON for every tool that had
-    /// `defer_loading: true` at option-parse time, keyed by tool name.
-    /// Populated by `apply_tool_search_client_injection` and later
-    /// drained by `AgentLoopState::new` when it builds the per-loop
-    /// client state. Never populated for native mode — the provider
-    /// handles deferral server-side.
-    pub deferred_bodies: std::collections::BTreeMap<String, serde_json::Value>,
-}
-
-impl ToolSearchConfig {
-    /// Default when the user writes `tool_search: true` with no detail.
-    pub(crate) fn default_bm25_auto() -> Self {
-        Self {
-            variant: ToolSearchVariant::Bm25,
-            mode: ToolSearchMode::Auto,
-            always_loaded: Vec::new(),
-            strategy: None,
-            budget_tokens: None,
-            name: None,
-            include_stub_listing: false,
-            deferred_bodies: std::collections::BTreeMap::new(),
-        }
-    }
-
-    /// Resolve the effective strategy, falling back to the variant
-    /// default when the user left `strategy` unset.
-    pub(crate) fn effective_strategy(&self) -> ToolSearchStrategy {
-        self.strategy
-            .unwrap_or_else(|| ToolSearchStrategy::default_for_variant(self.variant))
-    }
-
-    /// Resolve the synthetic tool's name. Default matches the spec's
-    /// proposed `__harn_tool_search` sentinel.
-    pub(crate) fn effective_name(&self) -> &str {
-        self.name.as_deref().unwrap_or("__harn_tool_search")
-    }
-}
-
-/// All options for an LLM API call, extracted once from user-facing args.
-#[derive(Clone)]
-pub(crate) struct LlmCallOptions {
-    // --- Routing ---
-    pub provider: String,
-    pub model: String,
-    pub api_key: String,
-
-    // --- Conversation ---
-    pub messages: Vec<serde_json::Value>,
-    pub system: Option<String>,
-    /// Optional short summary string prepended to the system prompt.
-    /// Populated by auto-compaction at mid-loop boundaries; callers
-    /// typically leave this `None`.
-    pub transcript_summary: Option<String>,
-
-    // --- Generation ---
-    pub max_tokens: i64,
-    pub temperature: Option<f64>,
-    pub top_p: Option<f64>,
-    pub top_k: Option<i64>,
-    pub stop: Option<Vec<String>>,
-    pub seed: Option<i64>,
-    pub frequency_penalty: Option<f64>,
-    pub presence_penalty: Option<f64>,
-
-    // --- Structured output ---
-    pub response_format: Option<String>,
-    pub json_schema: Option<serde_json::Value>,
-    pub output_schema: Option<serde_json::Value>,
-    pub output_validation: Option<String>,
-
-    // --- Thinking ---
-    pub thinking: Option<ThinkingConfig>,
-
-    // --- Tools ---
-    pub tools: Option<VmValue>,
-    pub native_tools: Option<Vec<serde_json::Value>>,
-    pub tool_choice: Option<serde_json::Value>,
-    /// Progressive-disclosure configuration. When set, the options
-    /// extractor resolves this against the active provider's capability
-    /// matrix and, for native-supporting providers, prepends a
-    /// `tool_search_tool_*_20251119` meta-tool to `native_tools`. For
-    /// client-executed mode (harn#70) this carries the config forward
-    /// into the agent-loop fallback. See [`ToolSearchConfig`].
-    #[allow(dead_code)] // consumed by the options extractor; persisted for transcript /
-    // replay fidelity and harn#70's client-executed loop
-    pub tool_search: Option<ToolSearchConfig>,
-
-    // --- Caching ---
-    pub cache: bool,
-
-    // --- Transport ---
-    pub timeout: Option<u64>,
-    /// Per-chunk idle timeout for streaming responses (seconds).
-    pub idle_timeout: Option<u64>,
-    /// When true, use streaming SSE transport (token-by-token deltas).
-    /// When false, use synchronous request/response. Default: true.
-    pub stream: bool,
-
-    // --- Provider-specific overrides ---
-    pub provider_overrides: Option<serde_json::Value>,
-
-    // --- Assistant prefill ---
-    /// Optional prefill string. When set, providers append a final
-    /// `role: "assistant"` message with this content so the model
-    /// continues from there. Cleared by the agent loop after each turn.
-    /// See `llm::providers::anthropic` and `llm::providers::openai_compat`
-    /// for provider-specific plumbing.
-    pub prefill: Option<String>,
-}
-
-/// Resolve effective request timeout: explicit value > `HARN_LLM_TIMEOUT` env > 120s default.
-fn resolve_timeout(explicit: Option<u64>) -> u64 {
-    explicit.unwrap_or_else(|| {
-        std::env::var("HARN_LLM_TIMEOUT")
-            .ok()
-            .and_then(|v| v.parse::<u64>().ok())
-            .unwrap_or(120)
-    })
-}
-
-impl LlmCallOptions {
-    pub(crate) fn resolve_timeout(&self) -> u64 {
-        resolve_timeout(self.timeout)
-    }
-}
-
-/// Send-safe subset of `LlmCallOptions` used for provider transport.
-#[derive(Clone, Debug)]
-pub(crate) struct LlmRequestPayload {
-    pub provider: String,
-    pub model: String,
-    pub api_key: String,
-    pub messages: Vec<serde_json::Value>,
-    pub system: Option<String>,
-    pub max_tokens: i64,
-    pub temperature: Option<f64>,
-    pub top_p: Option<f64>,
-    pub top_k: Option<i64>,
-    pub stop: Option<Vec<String>>,
-    pub seed: Option<i64>,
-    pub frequency_penalty: Option<f64>,
-    pub presence_penalty: Option<f64>,
-    pub response_format: Option<String>,
-    pub json_schema: Option<serde_json::Value>,
-    pub thinking: Option<ThinkingConfig>,
-    pub native_tools: Option<Vec<serde_json::Value>>,
-    pub tool_choice: Option<serde_json::Value>,
-    pub cache: bool,
-    pub timeout: Option<u64>,
-    pub stream: bool,
-    pub provider_overrides: Option<serde_json::Value>,
-    pub prefill: Option<String>,
-}
-
-impl LlmRequestPayload {
-    pub(crate) fn resolve_timeout(&self) -> u64 {
-        resolve_timeout(self.timeout)
-    }
-}
-
-impl From<&LlmCallOptions> for LlmRequestPayload {
-    fn from(opts: &LlmCallOptions) -> Self {
-        Self {
-            provider: opts.provider.clone(),
-            model: opts.model.clone(),
-            api_key: opts.api_key.clone(),
-            messages: opts.messages.clone(),
-            system: opts.system.clone(),
-            max_tokens: opts.max_tokens,
-            temperature: opts.temperature,
-            top_p: opts.top_p,
-            top_k: opts.top_k,
-            stop: opts.stop.clone(),
-            seed: opts.seed,
-            frequency_penalty: opts.frequency_penalty,
-            presence_penalty: opts.presence_penalty,
-            response_format: opts.response_format.clone(),
-            json_schema: opts.json_schema.clone(),
-            thinking: opts.thinking.clone(),
-            native_tools: opts.native_tools.clone(),
-            tool_choice: opts.tool_choice.clone(),
-            cache: opts.cache,
-            timeout: opts.timeout,
-            stream: opts.stream,
-            provider_overrides: opts.provider_overrides.clone(),
-            prefill: opts.prefill.clone(),
-        }
-    }
-}
-
-pub(crate) struct LlmResult {
-    pub text: String,
-    pub tool_calls: Vec<serde_json::Value>,
-    pub input_tokens: i64,
-    pub output_tokens: i64,
-    /// Prompt tokens served from the provider's cache (when supported).
-    /// Anthropic: `usage.cache_read_input_tokens`.
-    /// OpenAI: `usage.prompt_tokens_details.cached_tokens`.
-    /// OpenRouter passthrough for Anthropic: `usage.cache_read_input_tokens`.
-    /// Defaults to 0 when the provider doesn't report it.
-    pub cache_read_tokens: i64,
-    /// Prompt tokens written to the provider's cache on this request
-    /// (Anthropic `usage.cache_creation_input_tokens`). Helps distinguish
-    /// "warm-up" calls from cache hits.
-    pub cache_write_tokens: i64,
-    pub model: String,
-    pub provider: String,
-    pub thinking: Option<String>,
-    pub stop_reason: Option<String>,
-    pub blocks: Vec<serde_json::Value>,
-}
-
-pub(crate) fn vm_build_llm_result(
-    result: &LlmResult,
-    parsed_json: Option<VmValue>,
-    transcript: Option<VmValue>,
-    tools_val: Option<&VmValue>,
-) -> VmValue {
-    use crate::stdlib::json_to_vm_value;
-
-    let mut dict = BTreeMap::new();
-    dict.insert(
-        "text".to_string(),
-        VmValue::String(Rc::from(result.text.as_str())),
-    );
-    dict.insert(
-        "model".to_string(),
-        VmValue::String(Rc::from(result.model.as_str())),
-    );
-    dict.insert(
-        "provider".to_string(),
-        VmValue::String(Rc::from(result.provider.as_str())),
-    );
-    dict.insert(
-        "input_tokens".to_string(),
-        VmValue::Int(result.input_tokens),
-    );
-    dict.insert(
-        "output_tokens".to_string(),
-        VmValue::Int(result.output_tokens),
-    );
-    // Cache accounting (0 when provider doesn't report cache info).
-    dict.insert(
-        "cache_read_tokens".to_string(),
-        VmValue::Int(result.cache_read_tokens),
-    );
-    dict.insert(
-        "cache_write_tokens".to_string(),
-        VmValue::Int(result.cache_write_tokens),
-    );
-
-    if let Some(json_val) = parsed_json {
-        dict.insert("data".to_string(), json_val);
-    }
-
-    // Always run the tagged-protocol parser so canonical_text / violations /
-    // done_marker are available even with no tool registry. It's cheap and a
-    // no-op when the text has no tags. Native provider tool calls are
-    // authoritative and take precedence over text-parsed calls.
-    let tagged = Some(super::tools::parse_text_tool_calls_with_tools(
-        &result.text,
-        tools_val,
-    ));
-
-    let merged_tool_calls: Vec<serde_json::Value> = if !result.tool_calls.is_empty() {
-        result.tool_calls.clone()
-    } else if let Some(parse) = tagged.as_ref() {
-        parse.calls.clone()
-    } else {
-        Vec::new()
-    };
-    if !merged_tool_calls.is_empty() {
-        let calls: Vec<VmValue> = merged_tool_calls.iter().map(json_to_vm_value).collect();
-        dict.insert("tool_calls".to_string(), VmValue::List(Rc::new(calls)));
-    }
-
-    if let Some(parse) = tagged.as_ref() {
-        if !parse.violations.is_empty() {
-            let violations: Vec<VmValue> = parse
-                .violations
-                .iter()
-                .map(|v| VmValue::String(Rc::from(v.as_str())))
-                .collect();
-            dict.insert(
-                "protocol_violations".to_string(),
-                VmValue::List(Rc::new(violations)),
-            );
-        }
-        if !parse.errors.is_empty() {
-            let errors: Vec<VmValue> = parse
-                .errors
-                .iter()
-                .map(|e| VmValue::String(Rc::from(e.as_str())))
-                .collect();
-            dict.insert(
-                "tool_parse_errors".to_string(),
-                VmValue::List(Rc::new(errors)),
-            );
-        }
-        if let Some(ref body) = parse.done_marker {
-            dict.insert(
-                "done_marker".to_string(),
-                VmValue::String(Rc::from(body.as_str())),
-            );
-        }
-        if !parse.canonical.is_empty() {
-            dict.insert(
-                "canonical_text".to_string(),
-                VmValue::String(Rc::from(parse.canonical.as_str())),
-            );
-        }
-        // Always emit `prose` (fall back to raw text) so callers have a
-        // single reliable "the answer" key regardless of whether the model
-        // used the tagged protocol.
-        let prose = if parse.prose.is_empty() {
-            result.text.clone()
-        } else {
-            parse.prose.clone()
-        };
-        dict.insert(
-            "prose".to_string(),
-            VmValue::String(Rc::from(prose.as_str())),
-        );
-    }
-
-    if let Some(ref thinking) = result.thinking {
-        dict.insert(
-            "thinking".to_string(),
-            VmValue::String(Rc::from(thinking.as_str())),
-        );
-        dict.insert(
-            "private_reasoning".to_string(),
-            VmValue::String(Rc::from(thinking.as_str())),
-        );
-    }
-
-    if let Some(ref stop_reason) = result.stop_reason {
-        dict.insert(
-            "stop_reason".to_string(),
-            VmValue::String(Rc::from(stop_reason.as_str())),
-        );
-    }
-
-    if let Some(transcript) = transcript {
-        dict.insert("transcript".to_string(), transcript);
-    }
-
-    // Prose with fenceless TS tool-call expressions stripped. Agent_loop
-    // applies the same semantics on its final iteration.
-    let visible_text = if tools_val.is_some() && result.tool_calls.is_empty() {
-        let parse_result = super::tools::parse_text_tool_calls_with_tools(&result.text, tools_val);
-        parse_result.prose
-    } else {
-        result.text.clone()
-    };
-    dict.insert(
-        "visible_text".to_string(),
-        VmValue::String(Rc::from(visible_text.as_str())),
-    );
-    dict.insert(
-        "blocks".to_string(),
-        VmValue::List(Rc::new(
-            result
-                .blocks
-                .iter()
-                .map(json_to_vm_value)
-                .collect::<Vec<_>>(),
-        )),
-    );
-
-    VmValue::Dict(Rc::new(dict))
-}
+// ─── Chat entry points ──────────────────────────────────────────────────
 
 /// Execute an LLM call. Always goes through the streaming path with a
 /// discarding receiver so all callers share one code path for status/error
@@ -556,26 +76,6 @@ pub(crate) async fn vm_call_llm_full_streaming_offthread(
         ))))
     })?
     .map_err(|message| VmError::Thrown(VmValue::String(Rc::from(message))))
-}
-
-/// Execute a text completion / fill-in-the-middle call owned by Harn.
-pub(crate) async fn vm_call_completion_full(
-    opts: &LlmCallOptions,
-    prefix: &str,
-    suffix: Option<&str>,
-) -> Result<LlmResult, VmError> {
-    if opts.provider == "mock" {
-        return Ok(mock_completion_response(prefix, suffix));
-    }
-
-    let resolved = crate::llm_config::provider_config(&opts.provider);
-    let completion_endpoint = resolved.and_then(|p| p.completion_endpoint.as_deref());
-
-    match completion_endpoint {
-        Some("/api/generate") => vm_call_completion_ollama(opts, prefix, suffix).await,
-        Some(_) => vm_call_completion_openai_style(opts, prefix, suffix).await,
-        None => vm_call_completion_fallback(opts, prefix, suffix).await,
-    }
 }
 
 async fn vm_call_llm_full_inner(
@@ -704,2095 +204,11 @@ async fn try_fallback_provider(
         .map_err(|_| primary_message)
 }
 
-fn mock_completion_response(prefix: &str, suffix: Option<&str>) -> LlmResult {
-    let suffix = suffix.unwrap_or_default();
-    let text = format!(
-        "Mock completion after {} chars{}",
-        prefix.chars().count(),
-        if suffix.is_empty() {
-            String::new()
-        } else {
-            format!(" before {} chars", suffix.chars().count())
-        }
-    );
-    LlmResult {
-        text: text.clone(),
-        tool_calls: Vec::new(),
-        input_tokens: (prefix.len() + suffix.len()) as i64,
-        output_tokens: 16,
-        cache_read_tokens: 0,
-        cache_write_tokens: 0,
-        model: "mock".to_string(),
-        provider: "mock".to_string(),
-        thinking: None,
-        stop_reason: Some("stop".to_string()),
-        blocks: vec![serde_json::json!({
-            "type": "output_text",
-            "text": text,
-            "visibility": "public",
-        })],
-    }
-}
-
-async fn vm_call_completion_openai_style(
-    opts: &LlmCallOptions,
-    prefix: &str,
-    suffix: Option<&str>,
-) -> Result<LlmResult, VmError> {
-    let llm_timeout = opts.resolve_timeout();
-    let client = super::shared_blocking_client().clone();
-
-    let pdef = crate::llm_config::provider_config(&opts.provider);
-    let base_url = pdef
-        .map(crate::llm_config::resolve_base_url)
-        .unwrap_or_else(|| "https://api.openai.com/v1".to_string());
-    let endpoint = pdef
-        .and_then(|p| p.completion_endpoint.as_deref())
-        .unwrap_or("/completions");
-
-    let mut body = serde_json::json!({
-        "model": opts.model,
-        "prompt": prefix,
-        "max_tokens": opts.max_tokens,
-    });
-    if let Some(suffix) = suffix.filter(|s| !s.is_empty()) {
-        body["suffix"] = serde_json::json!(suffix);
-    }
-    if let Some(temp) = opts.temperature {
-        body["temperature"] = serde_json::json!(temp);
-    }
-    if let Some(top_p) = opts.top_p {
-        body["top_p"] = serde_json::json!(top_p);
-    }
-    if let Some(stop) = &opts.stop {
-        body["stop"] = serde_json::json!(stop);
-    }
-    if let Some(seed) = opts.seed {
-        body["seed"] = serde_json::json!(seed);
-    }
-    if let Some(overrides) = &opts.provider_overrides {
-        if let Some(obj) = overrides.as_object() {
-            for (k, v) in obj {
-                body[k] = v.clone();
-            }
-        }
-    }
-
-    let req = client
-        .post(format!("{base_url}{endpoint}"))
-        .header("Content-Type", "application/json")
-        .timeout(std::time::Duration::from_secs(llm_timeout))
-        .json(&body);
-    let req = apply_auth_headers(req, &opts.api_key, pdef);
-
-    let response = req.send().await.map_err(|e| {
-        VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{} completion API error: {e}",
-            opts.provider
-        ))))
-    })?;
-
-    let json: serde_json::Value = response.json().await.map_err(|e| {
-        VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{} completion response parse error: {e}",
-            opts.provider
-        ))))
-    })?;
-
-    if let Some(err) = json["error"]["message"].as_str() {
-        return Err(VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{} completion API error: {err}",
-            opts.provider
-        )))));
-    }
-
-    Ok(LlmResult {
-        text: json["choices"][0]["text"]
-            .as_str()
-            .unwrap_or("")
-            .to_string(),
-        tool_calls: Vec::new(),
-        input_tokens: json["usage"]["prompt_tokens"].as_i64().unwrap_or(0),
-        output_tokens: json["usage"]["completion_tokens"].as_i64().unwrap_or(0),
-        cache_read_tokens: extract_cache_read_tokens(&json["usage"]),
-        cache_write_tokens: extract_cache_write_tokens(&json["usage"]),
-        model: opts.model.clone(),
-        provider: opts.provider.clone(),
-        thinking: None,
-        stop_reason: json["choices"][0]["finish_reason"]
-            .as_str()
-            .map(|s| s.to_string()),
-        blocks: vec![serde_json::json!({
-            "type": "output_text",
-            "text": json["choices"][0]["text"].as_str().unwrap_or(""),
-            "visibility": "public",
-        })],
-    })
-}
-
-async fn vm_call_completion_ollama(
-    opts: &LlmCallOptions,
-    prefix: &str,
-    suffix: Option<&str>,
-) -> Result<LlmResult, VmError> {
-    let llm_timeout = opts.resolve_timeout();
-    let client = super::shared_blocking_client().clone();
-    let pdef = crate::llm_config::provider_config(&opts.provider);
-    let base_url = pdef
-        .map(crate::llm_config::resolve_base_url)
-        .unwrap_or_else(|| "http://localhost:11434".to_string());
-    let endpoint = pdef
-        .and_then(|p| p.completion_endpoint.as_deref())
-        .unwrap_or("/api/generate");
-
-    let mut options = serde_json::Map::new();
-    if let Some(num_ctx) = ollama_num_ctx_override() {
-        options.insert("num_ctx".to_string(), serde_json::json!(num_ctx));
-    }
-    if let Some(temp) = opts.temperature {
-        options.insert("temperature".to_string(), serde_json::json!(temp));
-    }
-    if let Some(top_p) = opts.top_p {
-        options.insert("top_p".to_string(), serde_json::json!(top_p));
-    }
-    if let Some(top_k) = opts.top_k {
-        options.insert("top_k".to_string(), serde_json::json!(top_k));
-    }
-    if let Some(seed) = opts.seed {
-        options.insert("seed".to_string(), serde_json::json!(seed));
-    }
-    if let Some(stop) = &opts.stop {
-        options.insert("stop".to_string(), serde_json::json!(stop));
-    }
-    options.insert(
-        "num_predict".to_string(),
-        serde_json::json!(opts.max_tokens),
-    );
-
-    let mut body = serde_json::json!({
-        "model": opts.model,
-        "prompt": prefix,
-        "stream": false,
-        "raw": true,
-        "options": options,
-    });
-    if let Some(suffix) = suffix.filter(|s| !s.is_empty()) {
-        body["suffix"] = serde_json::json!(suffix);
-    }
-    if let Some(system) = &opts.system {
-        body["system"] = serde_json::json!(system);
-    }
-    if let Some(keep_alive) = ollama_keep_alive_override() {
-        body["keep_alive"] = keep_alive;
-    }
-    if let Some(overrides) = &opts.provider_overrides {
-        if let Some(obj) = overrides.as_object() {
-            for (k, v) in obj {
-                body[k] = v.clone();
-            }
-        }
-    }
-
-    let req = client
-        .post(format!("{base_url}{endpoint}"))
-        .header("Content-Type", "application/json")
-        .timeout(std::time::Duration::from_secs(llm_timeout))
-        .json(&body);
-    let req = apply_auth_headers(req, &opts.api_key, pdef);
-
-    let response = req.send().await.map_err(|e| {
-        VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{} completion API error: {e}",
-            opts.provider
-        ))))
-    })?;
-    let json: serde_json::Value = response.json().await.map_err(|e| {
-        VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{} completion response parse error: {e}",
-            opts.provider
-        ))))
-    })?;
-    if let Some(err) = json["error"].as_str() {
-        return Err(VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{} completion API error: {err}",
-            opts.provider
-        )))));
-    }
-
-    Ok(LlmResult {
-        text: json["response"].as_str().unwrap_or("").to_string(),
-        tool_calls: Vec::new(),
-        input_tokens: json["prompt_eval_count"].as_i64().unwrap_or(0),
-        output_tokens: json["eval_count"].as_i64().unwrap_or(0),
-        cache_read_tokens: 0,
-        cache_write_tokens: 0,
-        model: opts.model.clone(),
-        provider: opts.provider.clone(),
-        thinking: None,
-        stop_reason: json["done_reason"].as_str().map(|s| s.to_string()),
-        blocks: vec![serde_json::json!({
-            "type": "output_text",
-            "text": json["response"].as_str().unwrap_or(""),
-            "visibility": "public",
-        })],
-    })
-}
-
-async fn vm_call_completion_fallback(
-    opts: &LlmCallOptions,
-    prefix: &str,
-    suffix: Option<&str>,
-) -> Result<LlmResult, VmError> {
-    let mut fallback_opts = opts.clone();
-    let mut instruction = String::from(
-        "Continue the user's text. Return only the missing continuation with no commentary, fences, or quoting.",
-    );
-    if let Some(suffix) = suffix.filter(|s| !s.is_empty()) {
-        instruction.push_str("\nRespect the required suffix exactly and produce only the text that belongs between PREFIX and SUFFIX.");
-        fallback_opts.messages = vec![serde_json::json!({
-            "role": "user",
-            "content": format!("PREFIX:\n{prefix}\n\nSUFFIX:\n{suffix}\n\nReturn only the missing text between PREFIX and SUFFIX."),
-        })];
-    } else {
-        fallback_opts.messages = vec![serde_json::json!({
-            "role": "user",
-            "content": format!("PREFIX:\n{prefix}\n\nReturn only the next continuation text."),
-        })];
-    }
-    fallback_opts.system = match &opts.system {
-        Some(system) => Some(format!("{system}\n\n{instruction}")),
-        None => Some(instruction),
-    };
-    vm_call_llm_full(&fallback_opts).await
-}
-
-fn render_openai_message_content_as_text(content: &serde_json::Value) -> String {
-    match content {
-        serde_json::Value::String(text) => text.clone(),
-        serde_json::Value::Array(blocks) => {
-            let mut rendered = String::new();
-            for block in blocks {
-                let block_type = block.get("type").and_then(|v| v.as_str()).unwrap_or("");
-                match block_type {
-                    "text" | "output_text" => {
-                        if let Some(text) = block.get("text").and_then(|v| v.as_str()) {
-                            rendered.push_str(text);
-                        }
-                    }
-                    "tool_result" => {
-                        let content = block
-                            .get("content")
-                            .and_then(|v| v.as_str())
-                            .unwrap_or_default();
-                        if !rendered.is_empty() {
-                            rendered.push_str("\n\n");
-                        }
-                        rendered.push_str("[Result] ");
-                        rendered.push_str(content);
-                    }
-                    "reasoning" | "thinking" => {
-                        if let Some(text) = block
-                            .get("text")
-                            .and_then(|v| v.as_str())
-                            .or_else(|| block.get("thinking").and_then(|v| v.as_str()))
-                        {
-                            if !rendered.is_empty() {
-                                rendered.push('\n');
-                            }
-                            rendered.push_str(text);
-                        }
-                    }
-                    _ => {
-                        if let Some(text) = block.get("text").and_then(|v| v.as_str()) {
-                            if !rendered.is_empty() {
-                                rendered.push('\n');
-                            }
-                            rendered.push_str(text);
-                        } else if !block.is_null() {
-                            if !rendered.is_empty() {
-                                rendered.push('\n');
-                            }
-                            rendered.push_str(&block.to_string());
-                        }
-                    }
-                }
-            }
-            rendered
-        }
-        serde_json::Value::Null => String::new(),
-        other => other.to_string(),
-    }
-}
-
-fn extract_openai_message_field_as_text(
-    message: &serde_json::Value,
-    field_names: &[&str],
-) -> String {
-    let mut combined = String::new();
-    for field_name in field_names {
-        let field_text = message
-            .get(*field_name)
-            .map(render_openai_message_content_as_text)
-            .unwrap_or_default();
-        if field_text.trim().is_empty() {
-            continue;
-        }
-        if !combined.is_empty() {
-            combined.push('\n');
-        }
-        combined.push_str(field_text.trim());
-    }
-    combined
-}
-
-fn append_paragraph(target: &mut String, text: &str) {
-    if text.trim().is_empty() {
-        return;
-    }
-    if !target.is_empty() {
-        target.push('\n');
-    }
-    target.push_str(text.trim());
-}
-
-fn normalize_openai_message_text(message: &serde_json::Value) -> (String, String) {
-    let raw_text = extract_openai_message_field_as_text(message, &["content"]);
-    let reasoning_text =
-        extract_openai_message_field_as_text(message, &["reasoning", "reasoning_content"]);
-    // Qwen3/3.5 emit inline `<think>...</think>` when
-    // `chat_template_kwargs.enable_thinking` is set. Split them out so the
-    // agent loop doesn't treat reasoning as output or parse tool calls
-    // inside them.
-    let (mut text, inline_thinking) = split_openai_thinking_blocks(&raw_text);
-    let mut extracted_thinking = String::new();
-    append_paragraph(&mut extracted_thinking, &reasoning_text);
-    append_paragraph(&mut extracted_thinking, &inline_thinking);
-    if text.is_empty() && !extracted_thinking.is_empty() {
-        text = extracted_thinking.clone();
-    }
-    (text, extracted_thinking)
-}
-
-pub(crate) fn normalize_openai_style_messages(
-    messages: Vec<serde_json::Value>,
-    force_string_content: bool,
-) -> Vec<serde_json::Value> {
-    messages
-        .into_iter()
-        .map(|message| {
-            let Some(object) = message.as_object() else {
-                return message;
-            };
-            let mut normalized = object.clone();
-            if force_string_content {
-                let content = normalized
-                    .get("content")
-                    .cloned()
-                    .unwrap_or(serde_json::Value::String(String::new()));
-                normalized.insert(
-                    "content".to_string(),
-                    serde_json::Value::String(render_openai_message_content_as_text(&content)),
-                );
-            }
-            serde_json::Value::Object(normalized)
-        })
-        .collect()
-}
-
-fn should_debug_message_shapes() -> bool {
-    std::env::var("HARN_DEBUG_MESSAGE_SHAPES")
-        .map(|value| value != "0" && !value.eq_ignore_ascii_case("false"))
-        .unwrap_or(false)
-}
-
-pub(crate) fn debug_log_message_shapes(label: &str, messages: &[serde_json::Value]) {
-    if !should_debug_message_shapes() {
-        return;
-    }
-    let summary = messages
-        .iter()
-        .enumerate()
-        .map(|(idx, message)| {
-            let role = message
-                .get("role")
-                .and_then(|value| value.as_str())
-                .unwrap_or("?");
-            let content_kind = match message.get("content") {
-                Some(serde_json::Value::String(_)) => "string",
-                Some(serde_json::Value::Null) => "null",
-                Some(serde_json::Value::Array(_)) => "array",
-                Some(serde_json::Value::Object(_)) => "object",
-                Some(_) => "other",
-                None => "missing",
-            };
-            let has_tool_call_id = message.get("tool_call_id").is_some();
-            let tool_calls = message
-                .get("tool_calls")
-                .and_then(|value| value.as_array())
-                .map(|calls| calls.len())
-                .unwrap_or(0);
-            let has_reasoning = message
-                .get("reasoning")
-                .map(|value| !value.is_null())
-                .unwrap_or(false);
-            format!(
-                "#{idx}:{role}:content={content_kind}:tool_call_id={has_tool_call_id}:tool_calls={tool_calls}:reasoning={has_reasoning}"
-            )
-        })
-        .collect::<Vec<_>>()
-        .join(" | ");
-    crate::events::log_info("llm.message_shape", &format!("{label}: {summary}"));
-}
-
-/// Build a tagged, provider-prefixed error message from a non-2xx HTTP
-/// response so downstream agent loops can react (e.g. trigger compaction on
-/// `context_overflow`, back off on `rate_limited`, surface everything else as
-/// `http_error`). Shared by both streaming and non-streaming transports so
-/// the classification never drifts between them.
-pub(crate) fn classify_http_error(
-    provider: &str,
-    status: reqwest::StatusCode,
-    retry_after: Option<&str>,
-    body: &str,
-) -> String {
-    // Patterns cover vLLM, OpenAI, Anthropic, and most OpenAI-compatibles.
-    let body_lower = body.to_lowercase();
-    let is_context_overflow = body_lower.contains("maximum context length")
-        || body_lower.contains("context length")
-        || body_lower.contains("context_length_exceeded")
-        || body_lower.contains("prompt is too long")
-        || body_lower.contains("prompt_tokens_exceeded")
-        || body_lower.contains("this model's maximum context")
-        || body_lower.contains("exceeds the maximum")
-        || (body_lower.contains("max_tokens") && body_lower.contains("exceed"));
-    let tag = if is_context_overflow {
-        "context_overflow"
-    } else if status.as_u16() == 429 {
-        "rate_limited"
-    } else {
-        "http_error"
-    };
-    let mut msg = format!("{provider} HTTP {status} [{tag}]: {body}");
-    if let Some(ra) = retry_after {
-        msg.push_str(&format!(" (retry-after: {ra})"));
-    }
-    msg
-}
-
-/// Dispatch an LLM API call to the appropriate provider. This is the main
-/// entry point that routes to provider-specific implementations via the
-/// provider plugin architecture.
-///
-/// The dispatch order is:
-/// 1. Check the thread-local provider registry (populated by `register_default_providers`)
-/// 2. Fall back to config-based resolution (for dynamically-configured providers)
-/// 3. Use the legacy inline dispatch as a final fallback
-async fn vm_call_llm_api(
-    opts: &LlmRequestPayload,
-    delta_tx: Option<DeltaSender>,
-) -> Result<LlmResult, VmError> {
-    let provider = &opts.provider;
-
-    if super::provider::is_provider_registered(provider) {
-        return dispatch_to_registered_provider(opts, delta_tx).await;
-    }
-
-    // Fallback for unregistered providers: dispatch by API style.
-    let resolved = super::helpers::ResolvedProvider::resolve(provider);
-    let is_ollama = provider == "ollama" || resolved.endpoint.contains("/api/chat");
-    let is_anthropic = resolved.is_anthropic_style;
-
-    let body = if is_ollama {
-        super::providers::OllamaProvider::build_request_body(opts)
-    } else if is_anthropic {
-        super::providers::AnthropicProvider::build_request_body(opts)
-    } else {
-        super::providers::OpenAiCompatibleProvider::build_request_body(opts, false)
-    };
-
-    vm_call_llm_api_with_body(opts, delta_tx, body, is_anthropic, is_ollama).await
-}
-
-/// Dispatch to a registered provider by name.
-///
-/// Provider selection uses trait methods (`is_mock()`, `is_local()`,
-/// `is_anthropic_style()`) instead of string comparisons so that each
-/// provider owns its own dispatch semantics.
-async fn dispatch_to_registered_provider(
-    opts: &LlmRequestPayload,
-    delta_tx: Option<DeltaSender>,
-) -> Result<LlmResult, VmError> {
-    // Providers are zero-cost unit structs constructed inline to avoid
-    // RefCell-across-await conflicts on a shared registry.
-    let provider = &opts.provider;
-    let resolved = super::helpers::ResolvedProvider::resolve(provider);
-
-    let mock = super::providers::MockProvider;
-    if mock.is_mock() && provider == mock.name() {
-        return mock.chat_impl(opts, delta_tx).await;
-    }
-
-    let ollama = super::providers::OllamaProvider;
-    if (provider == ollama.name() || resolved.endpoint.contains("/api/chat")) && ollama.is_local() {
-        return ollama.chat_impl(opts, delta_tx).await;
-    }
-
-    if resolved.is_anthropic_style {
-        let anthropic = super::providers::AnthropicProvider;
-        return anthropic.chat_impl(opts, delta_tx).await;
-    }
-
-    super::providers::OpenAiCompatibleProvider::new(provider.to_string())
-        .chat_impl(opts, delta_tx)
-        .await
-}
-
-/// Execute an LLM API call with a pre-built request body. This is the shared
-/// transport layer used by all provider implementations. It handles:
-/// - Provider-specific overrides merging
-/// - Stream vs non-stream transport selection
-/// - HTTP error classification
-/// - SSE and NDJSON response parsing
-///
-/// Provider implementations call this after building their provider-specific
-/// request body via `build_request_body()`.
-pub(crate) async fn vm_call_llm_api_with_body(
-    opts: &LlmRequestPayload,
-    delta_tx: Option<DeltaSender>,
-    mut body: serde_json::Value,
-    is_anthropic_style: bool,
-    is_ollama: bool,
-) -> Result<LlmResult, VmError> {
-    let provider = &opts.provider;
-    let model = &opts.model;
-    let wants_streaming = delta_tx.is_some() && opts.stream;
-
-    let resolved = super::helpers::ResolvedProvider::resolve(provider);
-    let use_stream_transport = if is_ollama && !opts.stream {
-        crate::events::log_warn(
-            "llm",
-            "stream=false is not supported by Ollama, using streaming",
-        );
-        true
-    } else {
-        wants_streaming || is_ollama
-    };
-
-    if let Some(ref overrides) = opts.provider_overrides {
-        if let Some(obj) = overrides.as_object() {
-            for (k, v) in obj {
-                body[k] = v.clone();
-            }
-        }
-    }
-
-    if let Some(messages) = body.get("messages").and_then(|value| value.as_array()) {
-        debug_log_message_shapes(
-            &format!("outbound provider={provider} model={model}"),
-            messages,
-        );
-    }
-
-    if use_stream_transport {
-        body["stream"] = serde_json::json!(true);
-        // OpenAI-style: request usage in the final streaming chunk.
-        if !is_anthropic_style && !is_ollama {
-            body["stream_options"] = serde_json::json!({"include_usage": true});
-        }
-    }
-
-    let client = if use_stream_transport {
-        super::shared_streaming_client().clone()
-    } else {
-        super::shared_blocking_client().clone()
-    };
-
-    let req = client
-        .post(resolved.url())
-        .header("Content-Type", "application/json")
-        .timeout(std::time::Duration::from_secs(opts.resolve_timeout()))
-        .json(&body);
-    let req = resolved.apply_headers(req, &opts.api_key);
-
-    if use_stream_transport {
-        let tx = if let Some(tx) = delta_tx {
-            tx
-        } else {
-            let (tx, _rx) = tokio::sync::mpsc::unbounded_channel::<String>();
-            tx
-        };
-        let response = req.send().await.map_err(|e| {
-            let kind = if e.is_timeout() {
-                "timeout"
-            } else if e.is_connect() {
-                "connect"
-            } else if e.is_request() {
-                "request_build"
-            } else if e.is_body() {
-                "body"
-            } else {
-                "unknown"
-            };
-            // "unknown" uses Debug repr to surface the inner cause.
-            let detail = if kind == "unknown" {
-                format!("{provider} stream error ({kind}): {e:?}")
-            } else {
-                format!("{provider} stream error ({kind}): {e}")
-            };
-            VmError::Thrown(VmValue::String(Rc::from(detail)))
-        })?;
-        if !response.status().is_success() {
-            let status = response.status();
-            let retry_after = response
-                .headers()
-                .get("retry-after")
-                .and_then(|v| v.to_str().ok())
-                .map(|s| s.to_string());
-            let body = response.text().await.unwrap_or_default();
-            let msg = classify_http_error(provider, status, retry_after.as_deref(), &body);
-            return Err(VmError::Thrown(VmValue::String(Rc::from(msg))));
-        }
-        let ct = response
-            .headers()
-            .get("content-type")
-            .and_then(|v| v.to_str().ok())
-            .unwrap_or("");
-        if ct.contains("text/event-stream") {
-            return vm_call_llm_api_sse_from_response(response, model, &resolved, tx).await;
-        }
-        return vm_call_llm_api_ndjson_from_response(response, model, tx).await;
-    }
-
-    let response = req.send().await.map_err(|e| {
-        VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{provider} API error: {e}"
-        ))))
-    })?;
-
-    // Check HTTP status BEFORE parsing the body as LLM response, or error
-    // responses (e.g. vLLM "prompt too long" 400) silently become malformed
-    // parse results and the agent loop retries against the same bad context.
-    if !response.status().is_success() {
-        let status = response.status();
-        let retry_after = response
-            .headers()
-            .get("retry-after")
-            .and_then(|v| v.to_str().ok())
-            .map(|s| s.to_string());
-        let body = response.text().await.unwrap_or_default();
-        let msg = classify_http_error(provider, status, retry_after.as_deref(), &body);
-        return Err(VmError::Thrown(VmValue::String(Rc::from(msg))));
-    }
-
-    let json: serde_json::Value = response.json().await.map_err(|e| {
-        VmError::Thrown(VmValue::String(Rc::from(format!(
-            "{provider} response parse error: {e}"
-        ))))
-    })?;
-
-    parse_llm_response(&json, provider, model, &resolved)
-}
-
-/// Parse a complete (non-streaming) LLM JSON response into an `LlmResult`.
-fn parse_llm_response(
-    json: &serde_json::Value,
-    provider: &str,
-    model: &str,
-    resolved: &super::helpers::ResolvedProvider<'_>,
-) -> Result<LlmResult, VmError> {
-    if resolved.is_anthropic_style {
-        let mut text = String::new();
-        let mut thinking_text = String::new();
-        let mut tool_calls = Vec::new();
-        let mut blocks = Vec::new();
-
-        if let Some(content) = json["content"].as_array() {
-            for block in content {
-                match block["type"].as_str() {
-                    Some("text") => {
-                        if let Some(t) = block["text"].as_str() {
-                            text.push_str(t);
-                            blocks.push(serde_json::json!({"type": "output_text", "text": t, "visibility": "public"}));
-                        }
-                    }
-                    Some("thinking") => {
-                        if let Some(t) = block["thinking"].as_str() {
-                            thinking_text.push_str(t);
-                            blocks.push(serde_json::json!({"type": "reasoning", "text": t, "visibility": "private"}));
-                        }
-                    }
-                    Some("tool_use") => {
-                        let name = block["name"].as_str().unwrap_or("").to_string();
-                        let id = block["id"].as_str().unwrap_or("").to_string();
-                        let input = block["input"].clone();
-                        tool_calls.push(serde_json::json!({
-                            "id": id,
-                            "name": name,
-                            "arguments": input,
-                        }));
-                        blocks.push(serde_json::json!({
-                            "type": "tool_call",
-                            "id": block["id"].clone(),
-                            "name": block["name"].clone(),
-                            "arguments": block["input"].clone(),
-                            "visibility": "internal",
-                        }));
-                    }
-                    Some("server_tool_use") => {
-                        // Anthropic's server-side tool-search tool emits
-                        // a `server_tool_use` content block when it
-                        // queries. The model never sees this as a
-                        // dispatchable tool — Anthropic executes it for
-                        // us — so we record it for transcript/replay
-                        // fidelity but do NOT add it to `tool_calls`.
-                        blocks.push(serde_json::json!({
-                            "type": "tool_search_query",
-                            "id": block["id"].clone(),
-                            "name": block["name"].clone(),
-                            "query": block["input"].clone(),
-                            "visibility": "internal",
-                        }));
-                    }
-                    Some("tool_search_tool_result") => {
-                        // Server-side search results. Anthropic
-                        // auto-expands the referenced tools inline on
-                        // subsequent turns; we just record the event so
-                        // replay/eval can see which tools were promoted
-                        // and when.
-                        let references: Vec<serde_json::Value> = block["content"]
-                            ["tool_references"]
-                            .as_array()
-                            .cloned()
-                            .unwrap_or_default();
-                        blocks.push(serde_json::json!({
-                            "type": "tool_search_result",
-                            "tool_use_id": block["tool_use_id"].clone(),
-                            "tool_references": references,
-                            "visibility": "internal",
-                        }));
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        if text.is_empty() && tool_calls.is_empty() {
-            if let Some(err) = json["error"]["message"].as_str() {
-                return Err(VmError::Thrown(VmValue::String(Rc::from(format!(
-                    "{provider} API error: {err}"
-                )))));
-            }
-        }
-
-        let input_tokens = json["usage"]["input_tokens"].as_i64().unwrap_or(0);
-        let output_tokens = json["usage"]["output_tokens"].as_i64().unwrap_or(0);
-        let cache_read_tokens = extract_cache_read_tokens(&json["usage"]);
-        let cache_write_tokens = extract_cache_write_tokens(&json["usage"]);
-        let stop_reason = json["stop_reason"].as_str().map(|s| s.to_string());
-
-        Ok(LlmResult {
-            text,
-            tool_calls,
-            input_tokens,
-            output_tokens,
-            cache_read_tokens,
-            cache_write_tokens,
-            model: model.to_string(),
-            provider: provider.to_string(),
-            thinking: if thinking_text.is_empty() {
-                None
-            } else {
-                Some(thinking_text)
-            },
-            stop_reason,
-            blocks,
-        })
-    } else {
-        if let Some(err) = json["error"]["message"].as_str() {
-            return Err(VmError::Thrown(VmValue::String(Rc::from(format!(
-                "{provider} API error: {err}"
-            )))));
-        }
-
-        let message = &json["choices"][0]["message"];
-        let (text, extracted_thinking) = normalize_openai_message_text(message);
-        let mut blocks = if text.is_empty() {
-            Vec::new()
-        } else {
-            vec![serde_json::json!({"type": "output_text", "text": text, "visibility": "public"})]
-        };
-        if !extracted_thinking.is_empty() {
-            blocks.insert(
-                0,
-                serde_json::json!({
-                    "type": "reasoning",
-                    "text": extracted_thinking,
-                    "visibility": "private",
-                }),
-            );
-        }
-
-        let mut tool_calls = Vec::new();
-        if let Some(calls) = message["tool_calls"].as_array() {
-            for call in calls {
-                // OpenAI Responses-API tool_search (harn#71) emits
-                // `tool_search_call` blocks when the server-hosted
-                // search runs. These are NOT dispatchable tools — the
-                // server executes them for us — so we record the query
-                // as a transcript event and continue without touching
-                // tool_calls. `tool_search_output` blocks on the
-                // response carry server results and are recorded
-                // symmetrically.
-                let call_type = call["type"].as_str().unwrap_or("");
-                if call_type == "tool_search_call" {
-                    let id = call["id"].as_str().unwrap_or("").to_string();
-                    let query = call.get("query").cloned().unwrap_or_else(|| {
-                        call.get("input")
-                            .cloned()
-                            .unwrap_or(serde_json::Value::Null)
-                    });
-                    blocks.push(serde_json::json!({
-                        "type": "tool_search_query",
-                        "id": id,
-                        "name": "tool_search",
-                        "query": query,
-                        "visibility": "internal",
-                    }));
-                    continue;
-                }
-                if call_type == "tool_search_output" {
-                    let tool_use_id = call["call_id"]
-                        .as_str()
-                        .or_else(|| call["id"].as_str())
-                        .unwrap_or("")
-                        .to_string();
-                    let references = call["tool_references"]
-                        .as_array()
-                        .cloned()
-                        .unwrap_or_default();
-                    blocks.push(serde_json::json!({
-                        "type": "tool_search_result",
-                        "tool_use_id": tool_use_id,
-                        "tool_references": references,
-                        "visibility": "internal",
-                    }));
-                    continue;
-                }
-                let name = call["function"]["name"].as_str().unwrap_or("").to_string();
-                let args_str = call["function"]["arguments"].as_str().unwrap_or("{}");
-                let arguments: serde_json::Value = match serde_json::from_str(args_str) {
-                    Ok(v) => v,
-                    Err(e) => {
-                        serde_json::json!({
-                            "__parse_error": format!(
-                                "Could not parse tool arguments as JSON: {}. Raw input: {}",
-                                e,
-                                &args_str[..args_str.len().min(200)]
-                            )
-                        })
-                    }
-                };
-                let id = call["id"].as_str().unwrap_or("").to_string();
-                tool_calls.push(serde_json::json!({
-                    "id": id,
-                    "name": name,
-                    "arguments": arguments,
-                }));
-                blocks.push(serde_json::json!({
-                    "type": "tool_call",
-                    "id": call["id"].clone(),
-                    "name": call["function"]["name"].clone(),
-                    "arguments": arguments.clone(),
-                    "visibility": "internal",
-                }));
-            }
-        }
-
-        let input_tokens = json["usage"]["prompt_tokens"].as_i64().unwrap_or(0);
-        let output_tokens = json["usage"]["completion_tokens"].as_i64().unwrap_or(0);
-        let cache_read_tokens = extract_cache_read_tokens(&json["usage"]);
-        let cache_write_tokens = extract_cache_write_tokens(&json["usage"]);
-        let stop_reason = json["choices"][0]["finish_reason"]
-            .as_str()
-            .map(|s| s.to_string());
-
-        // OpenAI Responses-API `tool_search_call` / `tool_search_output`
-        // blocks (harn#71) are server-executed and get stripped from
-        // `tool_calls` during parsing; they show up only as transcript
-        // blocks. Count their presence as "did deliver something" so
-        // the empty-response error below doesn't trip when the
-        // server's response consisted entirely of a search
-        // query/result exchange.
-        let has_tool_search_block = blocks.iter().any(|b| {
-            matches!(
-                b.get("type").and_then(|v| v.as_str()),
-                Some("tool_search_query") | Some("tool_search_result")
-            )
-        });
-        if text.is_empty()
-            && extracted_thinking.is_empty()
-            && output_tokens > 0
-            && tool_calls.is_empty()
-            && !has_tool_search_block
-        {
-            return Err(VmError::Thrown(VmValue::String(Rc::from(format!(
-                "openai-compatible model {model} reported completion_tokens={output_tokens} but delivered no content, reasoning, or tool calls"
-            )))));
-        }
-
-        Ok(LlmResult {
-            text,
-            tool_calls,
-            input_tokens,
-            output_tokens,
-            cache_read_tokens,
-            cache_write_tokens,
-            model: model.to_string(),
-            provider: provider.to_string(),
-            thinking: if extracted_thinking.is_empty() {
-                None
-            } else {
-                Some(extracted_thinking)
-            },
-            stop_reason,
-            blocks,
-        })
-    }
-}
-
-/// Extract cache-read token count from a provider `usage` JSON value,
-/// covering Anthropic, OpenAI (and OpenAI-compatibles), and OpenRouter
-/// passthrough field shapes. Returns 0 when the provider doesn't report it.
-fn extract_cache_read_tokens(usage: &serde_json::Value) -> i64 {
-    // Anthropic / OpenRouter passthrough: usage.cache_read_input_tokens
-    if let Some(n) = usage
-        .get("cache_read_input_tokens")
-        .and_then(|v| v.as_i64())
-    {
-        return n;
-    }
-    // OpenAI (and vLLM/SGLang when configured): usage.prompt_tokens_details.cached_tokens
-    if let Some(n) = usage
-        .get("prompt_tokens_details")
-        .and_then(|d| d.get("cached_tokens"))
-        .and_then(|v| v.as_i64())
-    {
-        return n;
-    }
-    // OpenRouter variants: cache_read_tokens / cached_prompt_tokens.
-    if let Some(n) = usage.get("cache_read_tokens").and_then(|v| v.as_i64()) {
-        return n;
-    }
-    if let Some(n) = usage.get("cached_prompt_tokens").and_then(|v| v.as_i64()) {
-        return n;
-    }
-    0
-}
-
-/// Extract cache-write (creation) token count from a provider `usage` JSON.
-/// Currently only Anthropic reports this explicitly.
-fn extract_cache_write_tokens(usage: &serde_json::Value) -> i64 {
-    usage
-        .get("cache_creation_input_tokens")
-        .and_then(|v| v.as_i64())
-        .unwrap_or(0)
-}
-
-/// Consume an SSE streaming response from an already-sent request.
-/// Parses `data: {...}` lines from the response body.
-async fn vm_call_llm_api_sse_from_response(
-    response: reqwest::Response,
-    model: &str,
-    resolved: &super::helpers::ResolvedProvider<'_>,
-    delta_tx: DeltaSender,
-) -> Result<LlmResult, VmError> {
-    use tokio::io::AsyncBufReadExt;
-    use tokio_stream::StreamExt;
-
-    let stream = response.bytes_stream();
-    let reader = tokio::io::BufReader::new(tokio_util::io::StreamReader::new(
-        stream.map(|r| r.map_err(std::io::Error::other)),
-    ));
-    let mut lines = reader.lines();
-
-    let mut text = String::new();
-    let mut input_tokens: i64 = 0;
-    let mut output_tokens: i64 = 0;
-    let mut tool_calls: Vec<serde_json::Value> = Vec::new();
-    let mut blocks: Vec<serde_json::Value> = Vec::new();
-
-    struct ToolBlock {
-        id: String,
-        name: String,
-        input_json: String,
-    }
-    let mut current_tool: Option<ToolBlock> = None;
-    // Mirror structure for server-side tool-search queries: Anthropic
-    // streams the query JSON the same way as a regular tool_use, but we
-    // route it to a `tool_search_query` transcript event instead of the
-    // dispatchable `tool_calls` vector.
-    struct ServerToolBlock {
-        id: String,
-        name: String,
-        input_json: String,
-    }
-    let mut current_server_tool: Option<ServerToolBlock> = None;
-    let mut thinking_text = String::new();
-    let mut in_thinking_block = false;
-    let mut stop_reason: Option<String> = None;
-    let mut cache_read_tokens: i64 = 0;
-    let mut cache_write_tokens: i64 = 0;
-
-    let mut oai_tool_map: std::collections::HashMap<u64, (String, String, String)> =
-        std::collections::HashMap::new();
-    // Qwen3/3.5 via vLLM emit inline `<think>...</think>`. Strip these
-    // out of the visible delta stream so the tool-call parser / progress
-    // UI only see the real response.
-    let mut oai_thinking_splitter = ThinkingStreamSplitter::new();
-
-    while let Ok(Some(line)) = lines.next_line().await {
-        let data = if let Some(d) = line.strip_prefix("data: ") {
-            d
-        } else {
-            continue;
-        };
-        if data == "[DONE]" {
-            break;
-        }
-        let json: serde_json::Value = match serde_json::from_str(data) {
-            Ok(v) => v,
-            Err(_) => continue,
-        };
-
-        if resolved.is_anthropic_style {
-            match json["type"].as_str() {
-                Some("message_start") => {
-                    if let Some(n) = json["message"]["usage"]["input_tokens"].as_i64() {
-                        input_tokens = n;
-                    }
-                    let usage = &json["message"]["usage"];
-                    let cr = extract_cache_read_tokens(usage);
-                    if cr > 0 {
-                        cache_read_tokens = cr;
-                    }
-                    let cw = extract_cache_write_tokens(usage);
-                    if cw > 0 {
-                        cache_write_tokens = cw;
-                    }
-                }
-                Some("content_block_start") => {
-                    let block = &json["content_block"];
-                    match block["type"].as_str() {
-                        Some("tool_use") => {
-                            current_tool = Some(ToolBlock {
-                                id: block["id"].as_str().unwrap_or("").to_string(),
-                                name: block["name"].as_str().unwrap_or("").to_string(),
-                                input_json: String::new(),
-                            });
-                        }
-                        Some("server_tool_use") => {
-                            current_server_tool = Some(ServerToolBlock {
-                                id: block["id"].as_str().unwrap_or("").to_string(),
-                                name: block["name"].as_str().unwrap_or("").to_string(),
-                                input_json: String::new(),
-                            });
-                        }
-                        Some("tool_search_tool_result") => {
-                            // Non-streaming content: Anthropic embeds the
-                            // references directly in the block_start
-                            // payload. Record immediately — no deltas
-                            // follow for this block type.
-                            let refs: Vec<serde_json::Value> = block["content"]["tool_references"]
-                                .as_array()
-                                .cloned()
-                                .unwrap_or_default();
-                            blocks.push(serde_json::json!({
-                                "type": "tool_search_result",
-                                "tool_use_id": block["tool_use_id"].clone(),
-                                "tool_references": refs,
-                                "visibility": "internal",
-                            }));
-                        }
-                        Some("thinking") => {
-                            in_thinking_block = true;
-                        }
-                        _ => {}
-                    }
-                }
-                Some("content_block_delta") => {
-                    let delta = &json["delta"];
-                    match delta["type"].as_str() {
-                        Some("text_delta") => {
-                            if let Some(t) = delta["text"].as_str() {
-                                text.push_str(t);
-                                let _ = delta_tx.send(t.to_string());
-                                blocks.push(serde_json::json!({"type": "output_text", "text": t, "visibility": "public"}));
-                            }
-                        }
-                        Some("thinking_delta") => {
-                            if let Some(t) = delta["thinking"].as_str() {
-                                thinking_text.push_str(t);
-                                blocks.push(serde_json::json!({"type": "reasoning", "text": t, "visibility": "private"}));
-                            }
-                        }
-                        Some("input_json_delta") => {
-                            if let Some(ref mut tool) = current_tool {
-                                if let Some(j) = delta["partial_json"].as_str() {
-                                    tool.input_json.push_str(j);
-                                }
-                            } else if let Some(ref mut server_tool) = current_server_tool {
-                                if let Some(j) = delta["partial_json"].as_str() {
-                                    server_tool.input_json.push_str(j);
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-                Some("content_block_stop") => {
-                    if let Some(tool) = current_tool.take() {
-                        let args = serde_json::from_str::<serde_json::Value>(&tool.input_json)
-                            .unwrap_or(serde_json::Value::Object(Default::default()));
-                        tool_calls.push(serde_json::json!({
-                            "id": tool.id, "name": tool.name, "arguments": args,
-                        }));
-                        blocks.push(serde_json::json!({"type": "tool_call", "id": tool.id, "name": tool.name, "arguments": args, "visibility": "internal"}));
-                    } else if let Some(server_tool) = current_server_tool.take() {
-                        // Emit a `tool_search_query` transcript event —
-                        // not dispatchable, just observability.
-                        let query =
-                            serde_json::from_str::<serde_json::Value>(&server_tool.input_json)
-                                .unwrap_or(serde_json::Value::Object(Default::default()));
-                        blocks.push(serde_json::json!({
-                            "type": "tool_search_query",
-                            "id": server_tool.id,
-                            "name": server_tool.name,
-                            "query": query,
-                            "visibility": "internal",
-                        }));
-                    }
-                    in_thinking_block = false;
-                }
-                Some("message_delta") => {
-                    if let Some(n) = json["usage"]["output_tokens"].as_i64() {
-                        output_tokens = n;
-                    }
-                    let usage = &json["usage"];
-                    let cr = extract_cache_read_tokens(usage);
-                    if cr > 0 {
-                        cache_read_tokens = cr;
-                    }
-                    let cw = extract_cache_write_tokens(usage);
-                    if cw > 0 {
-                        cache_write_tokens = cw;
-                    }
-                    if let Some(sr) = json["delta"]["stop_reason"].as_str() {
-                        stop_reason = Some(sr.to_string());
-                    }
-                }
-                _ => {}
-            }
-        } else {
-            let choice = &json["choices"][0];
-            let delta = &choice["delta"];
-
-            if let Some(content) = delta["content"].as_str() {
-                let visible = oai_thinking_splitter.push(content);
-                if !visible.is_empty() {
-                    text.push_str(&visible);
-                    let _ = delta_tx.send(visible.clone());
-                    blocks.push(serde_json::json!({"type": "output_text", "text": visible, "visibility": "public"}));
-                }
-            }
-            let reasoning_delta =
-                extract_openai_message_field_as_text(delta, &["reasoning", "reasoning_content"]);
-            if !reasoning_delta.is_empty() {
-                append_paragraph(&mut thinking_text, &reasoning_delta);
-                blocks.push(serde_json::json!({"type": "reasoning", "text": reasoning_delta, "visibility": "private"}));
-            }
-
-            // Only capture finish_reason once; OpenRouter can send
-            // duplicates (qwen-code#2402) that truncate in-progress tool
-            // calls.
-            if stop_reason.is_none() {
-                if let Some(fr) = choice["finish_reason"].as_str() {
-                    stop_reason = Some(fr.to_string());
-                }
-            }
-
-            if let Some(tcs) = delta["tool_calls"].as_array() {
-                for tc in tcs {
-                    // OpenAI Responses-API server-side tool_search
-                    // (harn#71) streams as `tool_search_call` /
-                    // `tool_search_output` entries in the tool_calls
-                    // array. Record them as transcript events, never
-                    // as dispatchable calls.
-                    let tc_type = tc["type"].as_str().unwrap_or("");
-                    if tc_type == "tool_search_call" {
-                        let id = tc["id"].as_str().unwrap_or("").to_string();
-                        let query = tc.get("query").cloned().unwrap_or_else(|| {
-                            tc.get("input").cloned().unwrap_or(serde_json::Value::Null)
-                        });
-                        blocks.push(serde_json::json!({
-                            "type": "tool_search_query",
-                            "id": id,
-                            "name": "tool_search",
-                            "query": query,
-                            "visibility": "internal",
-                        }));
-                        continue;
-                    }
-                    if tc_type == "tool_search_output" {
-                        let tool_use_id = tc["call_id"]
-                            .as_str()
-                            .or_else(|| tc["id"].as_str())
-                            .unwrap_or("")
-                            .to_string();
-                        let references = tc["tool_references"]
-                            .as_array()
-                            .cloned()
-                            .unwrap_or_default();
-                        blocks.push(serde_json::json!({
-                            "type": "tool_search_result",
-                            "tool_use_id": tool_use_id,
-                            "tool_references": references,
-                            "visibility": "internal",
-                        }));
-                        continue;
-                    }
-                    let idx = tc["index"].as_u64().unwrap_or(0);
-                    let entry = oai_tool_map.entry(idx).or_insert_with(|| {
-                        let id = tc["id"].as_str().unwrap_or("").to_string();
-                        let name = tc["function"]["name"].as_str().unwrap_or("").to_string();
-                        (id, name, String::new())
-                    });
-                    if let Some(args) = tc["function"]["arguments"].as_str() {
-                        entry.2.push_str(args);
-                    }
-                }
-            }
-
-            if let Some(usage) = json.get("usage") {
-                if let Some(n) = usage["prompt_tokens"].as_i64() {
-                    input_tokens = n;
-                }
-                if let Some(n) = usage["completion_tokens"].as_i64() {
-                    output_tokens = n;
-                }
-                let cr = extract_cache_read_tokens(usage);
-                if cr > 0 {
-                    cache_read_tokens = cr;
-                }
-                let cw = extract_cache_write_tokens(usage);
-                if cw > 0 {
-                    cache_write_tokens = cw;
-                }
-            }
-        }
-    }
-
-    for (_, (id, name, args_str)) in oai_tool_map {
-        let args = serde_json::from_str::<serde_json::Value>(&args_str)
-            .unwrap_or(serde_json::Value::Object(Default::default()));
-        tool_calls.push(serde_json::json!({
-            "id": id, "name": name, "arguments": args,
-        }));
-        blocks.push(serde_json::json!({"type": "tool_call", "id": id, "name": name, "arguments": args, "visibility": "internal"}));
-    }
-
-    let final_visible = oai_thinking_splitter.flush();
-    if !final_visible.is_empty() {
-        text.push_str(&final_visible);
-        let _ = delta_tx.send(final_visible.clone());
-        blocks.push(serde_json::json!({"type": "output_text", "text": final_visible, "visibility": "public"}));
-    }
-    if !oai_thinking_splitter.thinking.is_empty() {
-        append_paragraph(&mut thinking_text, &oai_thinking_splitter.thinking);
-    }
-
-    let _ = in_thinking_block; // suppress unused warning
-
-    if text.is_empty() && !thinking_text.is_empty() {
-        text = thinking_text.clone();
-        blocks
-            .push(serde_json::json!({"type": "output_text", "text": text, "visibility": "public"}));
-    }
-    let has_tool_search_block = blocks.iter().any(|b| {
-        matches!(
-            b.get("type").and_then(|v| v.as_str()),
-            Some("tool_search_query") | Some("tool_search_result")
-        )
-    });
-    if text.is_empty()
-        && thinking_text.is_empty()
-        && output_tokens > 0
-        && tool_calls.is_empty()
-        && !has_tool_search_block
-    {
-        return Err(VmError::Thrown(VmValue::String(Rc::from(format!(
-            "openai-compatible model {model} reported completion_tokens={output_tokens} but delivered no content, reasoning, or tool calls"
-        )))));
-    }
-
-    Ok(LlmResult {
-        text,
-        tool_calls,
-        input_tokens,
-        output_tokens,
-        cache_read_tokens,
-        cache_write_tokens,
-        model: model.to_string(),
-        provider: if resolved.is_anthropic_style {
-            "anthropic".to_string()
-        } else {
-            "openai".to_string()
-        },
-        thinking: if thinking_text.is_empty() {
-            None
-        } else {
-            Some(thinking_text)
-        },
-        stop_reason,
-        blocks,
-    })
-}
-
-// OpenAI-compatible servers (vLLM, text-generation-inference, LocalAI,
-// llama.cpp server) expose `max_model_len` via `GET /v1/models`. Query it
-// once so auto-compaction thresholds match the real window instead of
-// assuming 80K and letting the server silently truncate older turns.
-
-use std::collections::HashMap as StdHashMap;
-use std::sync::{Mutex as StdMutex, OnceLock as StdOnceLock};
-
-type ContextWindowKey = (String, String);
-type ContextWindowCache = StdMutex<StdHashMap<ContextWindowKey, Option<usize>>>;
-
-fn context_window_cache() -> &'static ContextWindowCache {
-    static CACHE: StdOnceLock<ContextWindowCache> = StdOnceLock::new();
-    CACHE.get_or_init(|| StdMutex::new(StdHashMap::new()))
-}
-
-/// Fetch the server-reported maximum context length for a given model, if
-/// available. Caches results per (base_url, model_id) so we only pay the
-/// discovery cost once per session.
-///
-/// Returns `None` when the provider doesn't expose `/v1/models`, when the
-/// model isn't found in the response, or when the request fails for any
-/// reason — callers should fall back to their default threshold.
-pub async fn fetch_provider_max_context(
-    provider: &str,
-    model: &str,
-    api_key: &str,
-) -> Option<usize> {
-    let pdef = crate::llm_config::provider_config(provider);
-    let base_url = pdef
-        .map(crate::llm_config::resolve_base_url)
-        .unwrap_or_else(|| "https://api.openai.com/v1".to_string());
-    let cache_key = (base_url.clone(), model.to_string());
-
-    // Fast path: cached (may be Some(n) or a cached None meaning "we tried
-    // and it doesn't work for this provider, don't keep asking").
-    if let Ok(cache) = context_window_cache().lock() {
-        if let Some(value) = cache.get(&cache_key) {
-            return *value;
-        }
-    }
-
-    let fetched = fetch_provider_max_context_uncached(provider, model, api_key, &base_url).await;
-    if let Ok(mut cache) = context_window_cache().lock() {
-        cache.insert(cache_key, fetched);
-    }
-    fetched
-}
-
-/// Hardcoded context window sizes for well-known model families where the
-/// provider API doesn't expose this information (Anthropic, OpenAI).
-/// Returns `None` for unknown models — callers fall through to API discovery.
-fn known_model_context_window(model: &str) -> Option<usize> {
-    if model.starts_with("claude-") {
-        return Some(200_000);
-    }
-    if model.starts_with("gpt-4o") || model.starts_with("gpt-4.1") || model.starts_with("chatgpt-")
-    {
-        return Some(128_000);
-    }
-    if model.starts_with("gpt-4-turbo")
-        || model == "gpt-4-0125-preview"
-        || model == "gpt-4-1106-preview"
-    {
-        return Some(128_000);
-    }
-    if model.starts_with("gpt-4") {
-        return Some(8_192);
-    }
-    if model.starts_with("gpt-3.5-turbo") {
-        return Some(16_385);
-    }
-    if model.starts_with("o1") || model.starts_with("o3") || model.starts_with("o4") {
-        return Some(200_000);
-    }
-    if model.contains("gemini-2") || model.contains("gemini-1.5") {
-        return Some(1_000_000);
-    }
-    if model.contains("gemini") {
-        return Some(128_000);
-    }
-    None
-}
-
-/// Fetch context window from Ollama's `/api/show` endpoint.
-/// Returns the num_ctx from model parameters, or the default 2048 if not set.
-async fn fetch_ollama_context_window(model: &str, base_url: &str) -> Option<usize> {
-    let client = super::shared_utility_client();
-    let url = format!("{}/api/show", base_url.trim_end_matches('/'));
-    let body = serde_json::json!({"name": model});
-    // Ollama is typically local — tight per-request timeout so we fail
-    // fast when it isn't running.
-    let response = client
-        .post(&url)
-        .json(&body)
-        .timeout(std::time::Duration::from_secs(5))
-        .send()
-        .await
-        .ok()?;
-    if !response.status().is_success() {
-        return None;
-    }
-    let json: serde_json::Value = response.json().await.ok()?;
-    if let Some(n) = json
-        .pointer("/model_info/general.context_length")
-        .or_else(|| json.pointer("/model_info/context_length"))
-        .and_then(|v| v.as_u64())
-    {
-        return Some(n as usize);
-    }
-    // OLLAMA_NUM_CTX env override for a user-configured context window.
-    if let Ok(val) = std::env::var("OLLAMA_NUM_CTX") {
-        if let Ok(n) = val.parse::<usize>() {
-            return Some(n);
-        }
-    }
-    // Let caller use its default threshold (Ollama defaults vary 2048-8192).
-    None
-}
-
-/// Fetch context window from an OpenAI-compatible `/models` endpoint.
-async fn fetch_openai_compatible_context_window(
-    provider: &str,
-    model: &str,
-    api_key: &str,
-    base_url: &str,
-) -> Option<usize> {
-    let pdef = crate::llm_config::provider_config(provider);
-    let client = super::shared_utility_client();
-    let url = format!("{base_url}/models");
-    let req = client
-        .get(&url)
-        .header("Content-Type", "application/json")
-        .timeout(std::time::Duration::from_secs(10));
-    let req = apply_auth_headers(req, api_key, pdef);
-    let response = req.send().await.ok()?;
-    if !response.status().is_success() {
-        return None;
-    }
-    let json: serde_json::Value = response.json().await.ok()?;
-    let data = json.get("data").and_then(|d| d.as_array())?;
-    for entry in data {
-        let id = entry.get("id").and_then(|v| v.as_str()).unwrap_or("");
-        if id != model {
-            continue;
-        }
-        // vLLM: "max_model_len"
-        if let Some(n) = entry.get("max_model_len").and_then(|v| v.as_u64()) {
-            return Some(n as usize);
-        }
-        // Some servers: "context_length"
-        if let Some(n) = entry.get("context_length").and_then(|v| v.as_u64()) {
-            return Some(n as usize);
-        }
-        // Others: "max_context_length" / "n_ctx"
-        if let Some(n) = entry.get("max_context_length").and_then(|v| v.as_u64()) {
-            return Some(n as usize);
-        }
-        if let Some(n) = entry.get("n_ctx").and_then(|v| v.as_u64()) {
-            return Some(n as usize);
-        }
-        // OpenRouter: top_provider.context_length
-        if let Some(n) = entry
-            .get("top_provider")
-            .and_then(|tp| tp.get("context_length"))
-            .and_then(|v| v.as_u64())
-        {
-            return Some(n as usize);
-        }
-        break;
-    }
-    None
-}
-
-async fn fetch_provider_max_context_uncached(
-    provider: &str,
-    model: &str,
-    api_key: &str,
-    base_url: &str,
-) -> Option<usize> {
-    if let Some(n) = known_model_context_window(model) {
-        return Some(n);
-    }
-
-    if provider == "ollama" {
-        return fetch_ollama_context_window(model, base_url).await;
-    }
-
-    let is_openai_compatible = matches!(
-        provider,
-        "local"
-            | "openai"
-            | "vllm"
-            | "groq"
-            | "together"
-            | "openrouter"
-            | "deepinfra"
-            | "fireworks"
-            | "huggingface"
-    );
-    if is_openai_compatible {
-        return fetch_openai_compatible_context_window(provider, model, api_key, base_url).await;
-    }
-
-    None
-}
-
-/// Derive an effective auto-compact token threshold from the server-reported
-/// max context window, applying a safety margin so the compaction fires
-/// *before* the request actually overflows. Returns `None` if no server value
-/// is available — callers should keep their default.
-///
-/// The 0.75 safety factor is deliberate: the input messages alone aren't the
-/// whole request (the response also needs token budget), and our token
-/// estimator uses a rough chars/4 heuristic that under-counts ~15-25% on
-/// English code. 0.75 gives us headroom for both without wasting too much.
-pub(crate) fn effective_threshold_from_max_context(max_context: usize) -> usize {
-    let bounded = max_context.max(4_096);
-    (bounded * 3) / 4
-}
-
-/// Apply discovered context-window information to an auto-compact config.
-///
-/// Sets up two-tier compaction thresholds based on the model's actual context
-/// window:
-///   - Tier-1 threshold: stays at configured value unless it would overflow.
-///   - Tier-2 hard limit: set to 75% of max context (the real overflow boundary).
-///
-/// Only lowers thresholds — never raises above what the user explicitly set.
-pub(crate) async fn adapt_auto_compact_to_provider(
-    ac: &mut crate::orchestration::AutoCompactConfig,
-    user_specified_threshold: bool,
-    user_specified_hard_limit: bool,
-    provider: &str,
-    model: &str,
-    api_key: &str,
-) {
-    let Some(max_ctx) = fetch_provider_max_context(provider, model, api_key).await else {
-        return;
-    };
-    let effective = effective_threshold_from_max_context(max_ctx);
-
-    // Tier-2 hard limit comes from the actual context window.
-    if !user_specified_hard_limit {
-        ac.hard_limit_tokens = Some(effective);
-    } else if let Some(ref mut hl) = ac.hard_limit_tokens {
-        // Clamp user's hard limit down if it would overflow.
-        if *hl > effective {
-            *hl = effective;
-        }
-    }
-
-    // Tier-1: clamp to the hard limit (which would make tier-1 pointless)
-    // or to 65% of max context, whichever is lower.
-    if user_specified_threshold {
-        if ac.token_threshold > effective {
-            ac.token_threshold = effective;
-        }
-    } else {
-        let tier1_from_context = (max_ctx * 13) / 20; // 65%
-        if ac.token_threshold > tier1_from_context {
-            ac.token_threshold = tier1_from_context;
-        }
-    }
-}
-
-/// Split `<think>...</think>` blocks out of an OpenAI-compatible response
-/// text. Returns `(visible_text, thinking_text)`. Handles multiple thinking
-/// blocks, malformed/unclosed tags (best-effort), and preserves original
-/// whitespace in the visible portion.
-///
-/// Used for Qwen3/Qwen3.5 thinking via vLLM's `chat_template_kwargs.enable_thinking`.
-pub(crate) fn split_openai_thinking_blocks(raw: &str) -> (String, String) {
-    if !raw.contains("<think>") {
-        return (raw.to_string(), String::new());
-    }
-    let mut visible = String::new();
-    let mut thinking = String::new();
-    let mut rest = raw;
-    loop {
-        if let Some(start) = rest.find("<think>") {
-            visible.push_str(&rest[..start]);
-            let after_tag = &rest[start + "<think>".len()..];
-            if let Some(end) = after_tag.find("</think>") {
-                thinking.push_str(&after_tag[..end]);
-                rest = &after_tag[end + "</think>".len()..];
-            } else {
-                // Unclosed <think>: treat everything after as thinking.
-                thinking.push_str(after_tag);
-                break;
-            }
-        } else {
-            visible.push_str(rest);
-            break;
-        }
-    }
-    // Models emit `<think>...</think>\nActual`; strip the blank line.
-    let visible = visible.trim_start_matches('\n').to_string();
-    (visible, thinking.trim().to_string())
-}
-
-/// Incremental splitter for OpenAI-style streaming content that may contain
-/// `<think>...</think>` blocks. Buffers a small suffix to handle tags split
-/// across delta chunks. Only emits visible (non-thinking) content to the
-/// delta channel; accumulates thinking separately for the final result.
-#[derive(Default)]
-pub(crate) struct ThinkingStreamSplitter {
-    /// True while we're inside a `<think>` block.
-    in_thinking: bool,
-    /// Carryover characters from the last delta that might be the start of a
-    /// `<think>` or `</think>` tag. Never longer than `</think>`.len() - 1.
-    carry: String,
-    /// Accumulated thinking text (returned at the end).
-    pub thinking: String,
-}
-
-impl ThinkingStreamSplitter {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Feed a new delta chunk. Returns the visible portion to forward
-    /// downstream (may be empty if the entire chunk was part of a thinking
-    /// block or got held back as carry).
-    pub fn push(&mut self, delta: &str) -> String {
-        let combined = {
-            let mut s = std::mem::take(&mut self.carry);
-            s.push_str(delta);
-            s
-        };
-        let mut visible_out = String::new();
-        let mut cursor = 0usize;
-        let bytes = combined.as_bytes();
-        while cursor < bytes.len() {
-            if self.in_thinking {
-                // Look for </think> in the remainder.
-                if let Some(rel) = combined[cursor..].find("</think>") {
-                    self.thinking.push_str(&combined[cursor..cursor + rel]);
-                    cursor += rel + "</think>".len();
-                    self.in_thinking = false;
-                } else {
-                    // Hold back up to len("</think>")-1 chars as potential
-                    // split-tag carry; emit the rest into thinking.
-                    let hold = "</think>".len() - 1;
-                    let remaining = combined.len() - cursor;
-                    if remaining <= hold {
-                        self.carry.push_str(&combined[cursor..]);
-                    } else {
-                        let mut split = combined.len() - hold;
-                        while split > cursor && !combined.is_char_boundary(split) {
-                            split -= 1;
-                        }
-                        self.thinking.push_str(&combined[cursor..split]);
-                        self.carry.push_str(&combined[split..]);
-                    }
-                    return visible_out;
-                }
-            } else {
-                // Look for <think> in the remainder.
-                if let Some(rel) = combined[cursor..].find("<think>") {
-                    visible_out.push_str(&combined[cursor..cursor + rel]);
-                    cursor += rel + "<think>".len();
-                    self.in_thinking = true;
-                } else {
-                    // Hold back len("<think>")-1 chars as potential split-tag
-                    // carry; emit the rest as visible.
-                    let hold = "<think>".len() - 1;
-                    let remaining = combined.len() - cursor;
-                    if remaining <= hold {
-                        self.carry.push_str(&combined[cursor..]);
-                    } else {
-                        let mut split = combined.len() - hold;
-                        // Floor to char boundary to avoid slicing inside a
-                        // multi-byte UTF-8 codepoint.
-                        while split > cursor && !combined.is_char_boundary(split) {
-                            split -= 1;
-                        }
-                        visible_out.push_str(&combined[cursor..split]);
-                        self.carry.push_str(&combined[split..]);
-                    }
-                    return visible_out;
-                }
-            }
-        }
-        visible_out
-    }
-
-    /// Flush any remaining carry as visible or thinking, depending on state.
-    /// Called when the stream terminates.
-    pub fn flush(&mut self) -> String {
-        let rest = std::mem::take(&mut self.carry);
-        if self.in_thinking {
-            self.thinking.push_str(&rest);
-            String::new()
-        } else {
-            rest
-        }
-    }
-}
-
-/// Apply auth headers to a request based on provider config.
-pub(crate) fn apply_auth_headers(
-    req: reqwest::RequestBuilder,
-    api_key: &str,
-    pdef: Option<&crate::llm_config::ProviderDef>,
-) -> reqwest::RequestBuilder {
-    if api_key.is_empty() {
-        return req;
-    }
-    if let Some(p) = pdef {
-        match p.auth_style.as_str() {
-            "header" => {
-                let header_name = p.auth_header.as_deref().unwrap_or("x-api-key");
-                req.header(header_name, api_key)
-            }
-            "bearer" => req.header("Authorization", format!("Bearer {api_key}")),
-            "none" => req,
-            _ => req.header("Authorization", format!("Bearer {api_key}")),
-        }
-    } else {
-        req.header("Authorization", format!("Bearer {api_key}"))
-    }
-}
-
-/// Consume an NDJSON streaming response, forwarding text deltas to `delta_tx`
-/// while accumulating the full result.
-///
-/// Supports Ollama format (one JSON object per line):
-/// `{"message":{"role":"assistant","content":"Hi"},"done":false}`
-/// Final line has `"done":true` with token counts.
-///
-/// Also supports OpenAI-compatible NDJSON where each line is `data: {...}`.
-async fn vm_call_llm_api_ndjson_from_response(
-    response: reqwest::Response,
-    model: &str,
-    delta_tx: DeltaSender,
-) -> Result<LlmResult, VmError> {
-    use tokio::io::AsyncBufReadExt;
-    use tokio_stream::StreamExt;
-
-    let stream = response.bytes_stream();
-    let reader = tokio::io::BufReader::new(tokio_util::io::StreamReader::new(
-        stream.map(|r| r.map_err(std::io::Error::other)),
-    ));
-    let mut lines = reader.lines();
-
-    let mut text = String::new();
-    let mut input_tokens: i64 = 0;
-    let mut output_tokens: i64 = 0;
-    let mut result_model = model.to_string();
-
-    let mut thinking_text = String::new();
-    let mut blocks = Vec::new();
-
-    while let Ok(Some(line)) = lines.next_line().await {
-        if line.is_empty() {
-            continue;
-        }
-        let json: serde_json::Value = match serde_json::from_str(&line) {
-            Ok(v) => v,
-            Err(_) => continue,
-        };
-
-        // Ollama streams content and thinking as separate channels for
-        // reasoning-capable models (gemma3/4, qwen3, etc.); we always set
-        // `think: true` so thinking tokens aren't dropped.
-        let content = json["message"]["content"].as_str().unwrap_or("");
-        let thinking = json["message"]["thinking"].as_str().unwrap_or("");
-        if !content.is_empty() {
-            text.push_str(content);
-            let _ = delta_tx.send(content.to_string());
-            blocks.push(
-                serde_json::json!({"type": "output_text", "text": content, "visibility": "public"}),
-            );
-        } else if !thinking.is_empty() {
-            thinking_text.push_str(thinking);
-            let _ = delta_tx.send(thinking.to_string());
-            blocks.push(
-                serde_json::json!({"type": "reasoning", "text": thinking, "visibility": "private"}),
-            );
-        }
-
-        if let Some(m) = json["model"].as_str() {
-            result_model = m.to_string();
-        }
-
-        if json["done"].as_bool() == Some(true) {
-            if let Some(n) = json["prompt_eval_count"].as_i64() {
-                input_tokens = n;
-            }
-            if let Some(n) = json["eval_count"].as_i64() {
-                output_tokens = n;
-            }
-            break;
-        }
-    }
-
-    let thinking = if thinking_text.is_empty() {
-        None
-    } else {
-        Some(thinking_text.clone())
-    };
-    if text.is_empty() && !thinking_text.is_empty() {
-        text = thinking_text;
-    }
-
-    // Guard against upstream parser bugs reporting generated tokens with
-    // no visible content. Observed with `gemma4:26b` + ollama's
-    // server-side `PARSER gemma4` on tool-heavy prompts: eval_count is
-    // nonzero but every delta and the done chunk are empty strings.
-    // Silently returning empty text would make the agent loop burn
-    // iterations on a no-op.
-    if text.is_empty() && output_tokens > 0 {
-        return Err(VmError::Thrown(VmValue::String(Rc::from(format!(
-            "ollama model {model} reported eval_count={output_tokens} but delivered no content or thinking — likely a server-side parser bug; try a different model"
-        )))));
-    }
-
-    Ok(LlmResult {
-        text,
-        tool_calls: Vec::new(),
-        input_tokens,
-        output_tokens,
-        cache_read_tokens: 0,
-        cache_write_tokens: 0,
-        model: result_model,
-        provider: "ollama".to_string(),
-        thinking,
-        stop_reason: None,
-        blocks,
-    })
-}
-
-pub(crate) fn ollama_num_ctx_override() -> Option<u64> {
-    for key in [
-        "BURIN_OLLAMA_NUM_CTX",
-        "OLLAMA_CONTEXT_LENGTH",
-        "OLLAMA_NUM_CTX",
-    ] {
-        if let Ok(raw) = std::env::var(key) {
-            if let Ok(parsed) = raw.trim().parse::<u64>() {
-                if parsed > 0 {
-                    return Some(parsed);
-                }
-            }
-        }
-    }
-    None
-}
-
-pub(crate) fn ollama_keep_alive_override() -> Option<serde_json::Value> {
-    for key in ["BURIN_OLLAMA_KEEP_ALIVE", "OLLAMA_KEEP_ALIVE"] {
-        if let Ok(raw) = std::env::var(key) {
-            let trimmed = raw.trim();
-            if !trimmed.is_empty() {
-                return Some(match trimmed.to_ascii_lowercase().as_str() {
-                    "default" => serde_json::json!("30m"),
-                    "forever" | "infinite" | "-1" => serde_json::json!(-1),
-                    _ => {
-                        if let Ok(n) = trimmed.parse::<i64>() {
-                            serde_json::json!(n)
-                        } else {
-                            serde_json::json!(trimmed)
-                        }
-                    }
-                });
-            }
-        }
-    }
-    None
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{
-        classify_http_error, normalize_openai_message_text, ollama_keep_alive_override,
-        ollama_num_ctx_override, split_openai_thinking_blocks,
-        vm_call_llm_full_streaming_offthread, LlmCallOptions, LlmRequestPayload, ThinkingConfig,
-        ThinkingStreamSplitter,
-    };
-    use crate::value::VmValue;
-    use std::rc::Rc;
-
-    #[test]
-    fn thinking_split_no_tags_returns_original() {
-        let (visible, thinking) = split_openai_thinking_blocks("just a plain response");
-        assert_eq!(visible, "just a plain response");
-        assert_eq!(thinking, "");
-    }
-
-    #[test]
-    fn thinking_split_single_block() {
-        let raw = "<think>step by step reasoning</think>\nThe answer is 42.";
-        let (visible, thinking) = split_openai_thinking_blocks(raw);
-        assert_eq!(visible, "The answer is 42.");
-        assert_eq!(thinking, "step by step reasoning");
-    }
-
-    #[test]
-    fn thinking_split_multiple_blocks() {
-        let raw = "<think>first</think>hello <think>second</think>world";
-        let (visible, thinking) = split_openai_thinking_blocks(raw);
-        assert_eq!(visible, "hello world");
-        assert_eq!(
-            thinking,
-            "first\nsecond".replace('\n', "") /* joined with empty */
-        );
-        // Invariant: neither block text leaked into visible.
-        assert!(!visible.contains("first"));
-        assert!(!visible.contains("second"));
-    }
-
-    #[test]
-    fn thinking_split_unclosed_block_captures_remainder() {
-        let raw = "<think>reasoning with no closing tag and then text";
-        let (visible, thinking) = split_openai_thinking_blocks(raw);
-        assert_eq!(visible, "");
-        assert!(thinking.contains("reasoning with no closing tag"));
-    }
-
-    #[test]
-    fn thinking_stream_splitter_handles_clean_boundaries() {
-        let mut s = ThinkingStreamSplitter::new();
-        let v1 = s.push("<think>");
-        let v2 = s.push("reasoning");
-        let v3 = s.push("</think>");
-        let v4 = s.push("visible answer");
-        let tail = s.flush();
-        assert_eq!(v1, "");
-        assert_eq!(v2, "");
-        assert_eq!(v3, "");
-        let combined = format!("{}{}{}{}{}", v1, v2, v3, v4, tail);
-        assert_eq!(combined, "visible answer");
-        assert_eq!(s.thinking, "reasoning");
-    }
-
-    #[test]
-    fn thinking_stream_splitter_handles_split_tags() {
-        let mut s = ThinkingStreamSplitter::new();
-        let v1 = s.push("<thi");
-        let v2 = s.push("nk>inside</thi");
-        let v3 = s.push("nk>after");
-        let tail = s.flush();
-        let combined = format!("{}{}{}{}", v1, v2, v3, tail);
-        assert_eq!(combined, "after");
-        assert_eq!(s.thinking, "inside");
-    }
-
-    #[test]
-    fn thinking_stream_splitter_passthrough_without_tags() {
-        let mut s = ThinkingStreamSplitter::new();
-        let v1 = s.push("hello ");
-        let v2 = s.push("world");
-        let tail = s.flush();
-        let combined = format!("{}{}{}", v1, v2, tail);
-        assert_eq!(combined, "hello world");
-        assert_eq!(s.thinking, "");
-    }
-
-    #[test]
-    fn normalize_openai_message_text_uses_reasoning_when_content_missing() {
-        let message = serde_json::json!({
-            "reasoning": "hello from reasoning"
-        });
-        let (visible, thinking) = normalize_openai_message_text(&message);
-        assert_eq!(visible, "hello from reasoning");
-        assert_eq!(thinking, "hello from reasoning");
-    }
-
-    #[test]
-    fn normalize_openai_message_text_merges_reasoning_and_inline_think_blocks() {
-        let message = serde_json::json!({
-            "content": "<think>inline reasoning</think>visible answer",
-            "reasoning": "separate reasoning"
-        });
-        let (visible, thinking) = normalize_openai_message_text(&message);
-        assert_eq!(visible, "visible answer");
-        assert_eq!(thinking, "separate reasoning\ninline reasoning");
-    }
-
+    use super::options::base_opts;
+    use super::{vm_call_llm_full_streaming_offthread, LlmRequestPayload, ThinkingConfig};
     use crate::llm::env_lock;
-
-    fn base_opts(provider: &str) -> LlmCallOptions {
-        LlmCallOptions {
-            provider: provider.to_string(),
-            model: "test-model".to_string(),
-            api_key: String::new(),
-            messages: vec![serde_json::json!({"role": "user", "content": "hello"})],
-            system: None,
-            transcript_summary: Some("summary".to_string()),
-            max_tokens: 64,
-            temperature: Some(0.2),
-            top_p: Some(0.8),
-            top_k: Some(40),
-            stop: Some(vec!["STOP".to_string()]),
-            seed: Some(7),
-            frequency_penalty: Some(0.1),
-            presence_penalty: Some(0.2),
-            response_format: Some("json".to_string()),
-            json_schema: Some(serde_json::json!({"type": "object"})),
-            output_schema: Some(serde_json::json!({"type": "object"})),
-            output_validation: Some("error".to_string()),
-            thinking: None,
-            tools: Some(VmValue::String(Rc::from("vm-local-tools"))),
-            native_tools: Some(vec![
-                serde_json::json!({"type": "function", "function": {"name": "tool"}}),
-            ]),
-            tool_choice: Some(serde_json::json!({
-                "type": "function",
-                "function": {"name": "tool"}
-            })),
-            tool_search: None,
-            cache: true,
-            stream: true,
-            timeout: Some(5),
-            idle_timeout: None,
-            provider_overrides: Some(serde_json::json!({"custom_flag": true})),
-            prefill: None,
-        }
-    }
-
-    fn assert_send<T: Send>() {}
-
-    #[test]
-    fn request_payload_is_send_safe_and_drops_vm_local_fields() {
-        let payload = LlmRequestPayload::from(&base_opts("openai"));
-        assert_send::<LlmRequestPayload>();
-        assert_eq!(payload.provider, "openai");
-        assert_eq!(payload.model, "test-model");
-        assert!(payload.native_tools.is_some());
-        assert!(payload.tool_choice.is_some());
-        assert_eq!(
-            payload.provider_overrides,
-            Some(serde_json::json!({"custom_flag": true}))
-        );
-    }
 
     #[test]
     fn openai_compat_prefill_appends_assistant_and_sets_chat_template_kwargs() {
@@ -3085,33 +501,6 @@ mod tests {
     }
 
     #[test]
-    fn ollama_num_ctx_override_prefers_burin_env() {
-        let _guard = env_lock().lock().expect("env lock");
-        unsafe {
-            std::env::set_var("BURIN_OLLAMA_NUM_CTX", "131072");
-            std::env::remove_var("OLLAMA_CONTEXT_LENGTH");
-            std::env::remove_var("OLLAMA_NUM_CTX");
-        }
-        assert_eq!(ollama_num_ctx_override(), Some(131072));
-        unsafe {
-            std::env::remove_var("BURIN_OLLAMA_NUM_CTX");
-        }
-    }
-
-    #[test]
-    fn ollama_keep_alive_override_normalizes_forever() {
-        let _guard = env_lock().lock().expect("env lock");
-        unsafe {
-            std::env::set_var("BURIN_OLLAMA_KEEP_ALIVE", "forever");
-            std::env::remove_var("OLLAMA_KEEP_ALIVE");
-        }
-        assert_eq!(ollama_keep_alive_override(), Some(serde_json::json!(-1)));
-        unsafe {
-            std::env::remove_var("BURIN_OLLAMA_KEEP_ALIVE");
-        }
-    }
-
-    #[test]
     fn offthread_streaming_completes_inside_localset() {
         let _guard = env_lock().lock().expect("env lock");
         let runtime = tokio::runtime::Builder::new_multi_thread()
@@ -3224,67 +613,6 @@ mod tests {
             assert_eq!(json["keep_alive"].as_i64(), Some(-1));
             assert_eq!(json["options"]["num_ctx"].as_u64(), Some(131072));
         });
-    }
-
-    #[test]
-    fn classify_tags_vllm_prompt_too_long_as_context_overflow() {
-        let msg = classify_http_error(
-            "local",
-            reqwest::StatusCode::BAD_REQUEST,
-            None,
-            r#"{"object":"error","message":"This model's maximum context length is 8192 tokens. However, your prompt is too long (10234 tokens)."}"#,
-        );
-        assert!(msg.contains("[context_overflow]"), "msg was: {msg}");
-        assert!(msg.starts_with("local HTTP 400 Bad Request"));
-        assert!(!msg.contains("(retry-after"));
-    }
-
-    #[test]
-    fn classify_tags_openai_context_length_exceeded_as_context_overflow() {
-        let msg = classify_http_error(
-            "openai",
-            reqwest::StatusCode::BAD_REQUEST,
-            None,
-            r#"{"error":{"code":"context_length_exceeded","message":"maximum context length"}}"#,
-        );
-        assert!(msg.contains("[context_overflow]"), "msg was: {msg}");
-    }
-
-    #[test]
-    fn classify_tags_429_with_retry_after_as_rate_limited() {
-        let msg = classify_http_error(
-            "anthropic",
-            reqwest::StatusCode::TOO_MANY_REQUESTS,
-            Some("12"),
-            r#"{"error":{"type":"rate_limit_error","message":"quota exceeded"}}"#,
-        );
-        assert!(msg.contains("[rate_limited]"), "msg was: {msg}");
-        assert!(msg.ends_with("(retry-after: 12)"), "msg was: {msg}");
-    }
-
-    #[test]
-    fn classify_tags_opaque_500_as_http_error() {
-        let msg = classify_http_error(
-            "local",
-            reqwest::StatusCode::INTERNAL_SERVER_ERROR,
-            None,
-            "upstream exploded",
-        );
-        assert!(msg.contains("[http_error]"), "msg was: {msg}");
-        assert!(msg.contains("upstream exploded"));
-    }
-
-    #[test]
-    fn classify_429_with_context_body_still_prefers_context_overflow() {
-        // Some OpenAI-compat servers return 429 for context overflow;
-        // classify by body because caller reaction differs (compact vs back off).
-        let msg = classify_http_error(
-            "local",
-            reqwest::StatusCode::TOO_MANY_REQUESTS,
-            Some("1"),
-            "prompt is too long",
-        );
-        assert!(msg.contains("[context_overflow]"), "msg was: {msg}");
     }
 
     /// Bind a stub listener + spawn a responder that serves a single canned
@@ -3429,170 +757,5 @@ mod tests {
         );
         assert!(err.contains("[http_error]"), "err was: {err}");
         assert!(err.contains("upstream exploded"), "err was: {err}");
-    }
-
-    // ─── Tool Vault: server_tool_use + tool_search_tool_result parsing ──────
-
-    // Build a ResolvedProvider for the Anthropic path without going through
-    // the thread-local provider registry — these parser tests only need the
-    // is_anthropic_style flag set.
-    fn anthropic_resolved() -> super::super::helpers::ResolvedProvider<'static> {
-        super::super::helpers::ResolvedProvider::resolve("anthropic")
-    }
-
-    #[test]
-    fn anthropic_parser_records_server_tool_use_as_tool_search_query() {
-        let resolved = anthropic_resolved();
-        // Build a minimal Anthropic Messages API response containing a
-        // server_tool_use block (the model calling the search tool).
-        let response = serde_json::json!({
-            "content": [
-                {"type": "text", "text": "searching now"},
-                {
-                    "type": "server_tool_use",
-                    "id": "srvtoolu_01",
-                    "name": "tool_search_tool_bm25",
-                    "input": {"query": "weather"}
-                }
-            ],
-            "usage": {"input_tokens": 10, "output_tokens": 5}
-        });
-        let result =
-            super::parse_llm_response(&response, "anthropic", "claude-opus-4-7", &resolved)
-                .expect("parser succeeds");
-
-        // tool_calls is for *dispatchable* user tools — server-side tools
-        // must never appear there.
-        assert!(result.tool_calls.is_empty());
-
-        // The tool_search_query event is on the blocks list.
-        let has_query_event = result.blocks.iter().any(|b| {
-            b.get("type").and_then(|v| v.as_str()) == Some("tool_search_query")
-                && b.get("name").and_then(|v| v.as_str()) == Some("tool_search_tool_bm25")
-        });
-        assert!(
-            has_query_event,
-            "expected tool_search_query block; got {:#?}",
-            result.blocks
-        );
-    }
-
-    #[test]
-    fn openai_parser_records_tool_search_call_as_query_event() {
-        // OpenAI's Responses API (harn#71) surfaces the server-hosted
-        // tool_search as a `tool_search_call` entry in the `tool_calls`
-        // array. The parser must NOT add it to the dispatchable
-        // `tool_calls` vector — OpenAI runs the search on their side —
-        // but must record a `tool_search_query` transcript block so
-        // replay lines up with the Anthropic path.
-        let resolved = super::super::helpers::ResolvedProvider::resolve("openai");
-        let response = serde_json::json!({
-            "choices": [{
-                "message": {
-                    "content": "searching",
-                    "tool_calls": [
-                        {
-                            "id": "tsc_01",
-                            "type": "tool_search_call",
-                            "query": {"q": "weather"}
-                        }
-                    ]
-                },
-                "finish_reason": "tool_calls"
-            }],
-            "usage": {"prompt_tokens": 10, "completion_tokens": 5}
-        });
-        let result = super::parse_llm_response(&response, "openai", "gpt-5.4-preview", &resolved)
-            .expect("parser succeeds");
-
-        assert!(
-            result.tool_calls.is_empty(),
-            "tool_search_call is server-executed; must not be dispatchable"
-        );
-        let query = result
-            .blocks
-            .iter()
-            .find(|b| b.get("type").and_then(|v| v.as_str()) == Some("tool_search_query"))
-            .expect("tool_search_query block present");
-        assert_eq!(query["id"].as_str(), Some("tsc_01"));
-        assert_eq!(query["query"]["q"].as_str(), Some("weather"));
-    }
-
-    #[test]
-    fn openai_parser_records_tool_search_output_as_result_event() {
-        let resolved = super::super::helpers::ResolvedProvider::resolve("openai");
-        let response = serde_json::json!({
-            "choices": [{
-                "message": {
-                    "content": "",
-                    "tool_calls": [
-                        {
-                            "id": "tso_01",
-                            "type": "tool_search_output",
-                            "call_id": "tsc_01",
-                            "tool_references": [
-                                {"tool_name": "get_weather"}
-                            ]
-                        }
-                    ]
-                },
-                "finish_reason": "stop"
-            }],
-            "usage": {"prompt_tokens": 3, "completion_tokens": 1}
-        });
-        let result = super::parse_llm_response(&response, "openai", "gpt-5.4-preview", &resolved)
-            .expect("parser succeeds");
-
-        assert!(result.tool_calls.is_empty());
-        let result_block = result
-            .blocks
-            .iter()
-            .find(|b| b.get("type").and_then(|v| v.as_str()) == Some("tool_search_result"))
-            .expect("tool_search_result block present");
-        assert_eq!(result_block["tool_use_id"].as_str(), Some("tsc_01"));
-        let refs = result_block["tool_references"]
-            .as_array()
-            .expect("tool_references array");
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0]["tool_name"].as_str(), Some("get_weather"));
-    }
-
-    #[test]
-    fn anthropic_parser_records_tool_search_tool_result_as_event() {
-        let resolved = anthropic_resolved();
-        let response = serde_json::json!({
-            "content": [
-                {
-                    "type": "tool_search_tool_result",
-                    "tool_use_id": "srvtoolu_01",
-                    "content": {
-                        "type": "tool_search_tool_search_result",
-                        "tool_references": [
-                            {"type": "tool_reference", "tool_name": "get_weather"}
-                        ]
-                    }
-                },
-                {"type": "text", "text": "ok"}
-            ],
-            "usage": {"input_tokens": 3, "output_tokens": 1}
-        });
-        let result =
-            super::parse_llm_response(&response, "anthropic", "claude-opus-4-7", &resolved)
-                .expect("parser succeeds");
-
-        let result_block = result
-            .blocks
-            .iter()
-            .find(|b| b.get("type").and_then(|v| v.as_str()) == Some("tool_search_result"))
-            .expect("tool_search_result block present");
-        let refs = result_block["tool_references"]
-            .as_array()
-            .expect("tool_references array");
-        assert_eq!(refs.len(), 1);
-        assert_eq!(
-            refs[0]["tool_name"].as_str(),
-            Some("get_weather"),
-            "reference name preserved"
-        );
     }
 }
